EpiNow2: Estimate Real-Time Case Counts and Time-Varying Epidemiological Parameters

Estimates the time-varying reproduction number, rate of spread, and doubling time using a range of open-source tools (Abbott et al. (2020) <doi:10.12688/wellcomeopenres.16006.1>), and current best practices (Gostic et al. (2020) <doi:10.1101/2020.06.18.20134858>). It aims to help users avoid some of the limitations of naive implementations in a framework that is informed by community feedback and is under active development.

Author

Maintainer: Sam Abbott sam.abbott@lshtm.ac.uk (ORCID)

Authors:

Joel Hellewell joel.hellewell@lshtm.ac.uk (ORCID)

Robin Thompson robin.thompson@lshtm.ac.uk

Katelyn Gostic kgostic@uchicago.edu

Katharine Sherratt katharine.sherratt@lshtm.ac.uk

Hamada S. Badr badr@jhu.edu (ORCID)

Michael DeWitt me.dewitt.jr@gmail.com (ORCID)

Sophie Meakin sophie.meaking@lshtm.ac.uk

James Munday james.munday@lshtm.ac.uk

Nikos Bosse

Joe Hickson joseph.hickson@metoffice.gov.uk

EpiForecasts

Sebastian Funk sebastian.funk@lshtm.ac.uk

Other contributors:

Paul Mee paul.mee@lshtm.ac.uk [contributor]

Peter Ellis peter.ellis2013nz@gmail.com [contributor]