Delay Distribution Options

Returns delay distributions formatted for usage by downstream functions.

Usage

delay_opts(dist = dist_spec(), ..., fixed = FALSE)

Arguments

dist

A delay distribution or series of delay distributions generated using dist_spec(). Default is an empty call to dist_spec(), i.e. no delay

...

deprecated; use dist instead

fixed

deprecated; use dist instead

Value

A <delay_opts> object summarising the input delay distributions.

See also

Author

Sam Abbott

Sebastian Funk

Examples

# no delays
delay_opts()
#> 
#> Empty `dist_spec` distribution.

# A single delay that has uncertainty
delay <- dist_spec(mean = 1, mean_sd = 0.2, sd = 0.5, sd_sd = 0.1, max = 14)
delay_opts(delay)
#> 
#>   Uncertain lognormal distribution with (untruncated) logmean 1 (SD 0.2) and logSD 0.5 (SD 0.1)
#> 

# A single delay without uncertainty
delay <- dist_spec(mean = 1, sd = 0.5, max = 14)
delay_opts(delay)
#> 
#>   Fixed distribution with PMF [0.023 0.25 0.31 0.2 0.11 0.055 0.027 0.014 0.0071 0.0037 0.002 0.0011 0.00062 0.00035 2e-04]
#> 

# Multiple delays (in this case twice the same)
delay_opts(delay + delay)
#> 
#>   Fixed distribution with PMF [0.00052 0.011 0.075 0.16 0.2 0.18 0.14 0.092 0.058 0.035 0.021 0.012 0.0071 0.0042 0.0024 0.0014 0.00079]
#>