Estimate Infections, the Time-Varying Reproduction Number and the Rate of Growth

This function uses a non-parametric approach to reconstruct cases by date of infection from reported cases. It can optionally then estimate the time-varying reproduction number and the rate of growth.

estimate_infections(
  reported_cases,
  family = "negbin",
  incubation_period,
  reporting_delay,
  generation_time,
  gp = list(basis_prop = 0.5, boundary_scale = 2),
  rt_prior = list(mean = 1, sd = 1),
  prior_smoothing_window = 7,
  horizon = 7,
  model,
  cores = 1,
  chains = 2,
  samples = 1000,
  warmup = 200,
  estimate_rt = TRUE,
  adapt_delta = 0.99,
  max_treedepth = 15,
  return_fit = FALSE,
  verbose = TRUE,
  debug = FALSE
)

Arguments

reported_cases
A data frame of confirmed cases (confirm) by date (date).

family
A character string indicating the reporting model to use. Defaults to negative binomial ("negbin") with poisson ("poisson") also supported.

incubation_period
A list containing the mean, standard deviation of the mean (mean_sd), standard deviation (sd), standard deviation of the standard deviation and the maximum allowed value for the incubation period (assuming a lognormal distribution with all parameters excepting the max allowed value on the log scale).

reporting_delay
A list containing the mean, standard deviation of the mean (mean_sd), standard deviation (sd), standard deviation of the standard deviation and the maximum allowed value for the reporting delay (assuming a lognormal distribution with all parameters excepting the max allowed value on the log scale).

generation_time
A list containing the mean, standard deviation of the mean (mean_sd), standard deviation (sd), standard deviation of the standard deviation and the maximum allowed value for the generation time (assuming a gamma distribution).

gp
List controlling the Gaussian process approximation. Must contain the basis_prop (number of basis functions based on scaling the time points) which defaults to 0.5 and must be between 0 and 1 (increasing this increases the accuracy of the approximation and the cost of additional compute. Must also contain the boundary_scale (multiplied by half the range of the input time series). Increasing this increases the accuracy of the approximation at the cost of additional compute. See here: https://arxiv.org/abs/2004.11408 for more information on setting these parameters.

rt_prior
A list contain the mean and standard deviation (sd) of the gamma distributed prior for Rt. By default this is assumed to be mean 1 with a standard deviation of 1.

prior_smoothing_window
Numeric defaults to 7. The number of days over which to take a rolling average for the prior based on reported cases.

horizon
Numeric, defaults to 7. Number of days into the future to forecast.

model
A compiled stan model. By default uses the internal package model.

cores
Numeric, defaults to 2. The number of cores to use when fitting the stan model.

chains
Numeric, defaults to 2. The number of MCMC chains to use.

samples
Numeric, defaults to 1000. Number of samples post warmup.

warmup
Numeric, defaults to 200. Number of iteration of warmup to use.

estimate_rt
Logical, defaults TRUE. Should Rt be estimated when imputing infections.

adapt_delta
Numeric, defaults to 0.99. See ?rstan::sampling.

max_treedepth
Numeric, defaults to 15. See ?rstan::sampling.

return_fit
Logical, defaults to FALSE. Should the fitted stan model be returned.

verbose
Logical, defaults to TRUE. Should verbose progress messages be printed.

debug
Logical, defaults to FALSE. Enables debug model in which additional diagnostics are available

Examples

if (FALSE) {
## Get example case counts
reported_cases <- EpiNow2::example_confirmed[1:50]

## Set up example generation time
generation_time <- list(mean = EpiNow2::covid_generation_times[1, ]$mean,
                        mean_sd = EpiNow2::covid_generation_times[1, ]$mean_sd,
                        sd = EpiNow2::covid_generation_times[1, ]$sd,
                        sd_sd = EpiNow2::covid_generation_times[1, ]$sd_sd,
                        max = 30)
## Set                   
incubation_period <- list(mean = EpiNow2::covid_incubation_period[1, ]$mean,
                          mean_sd = EpiNow2::covid_incubation_period[1, ]$mean_sd,
                          sd = EpiNow2::covid_incubation_period[1, ]$sd,
                          sd_sd = EpiNow2::covid_incubation_period[1, ]$sd_sd,
                          max = 30)

reporting_delay <- list(mean = log(5),
                        mean_sd = log(2),
                        sd = log(2),
                        sd_sd = log(1.5),
                        max = 30)

## Run model
out <- estimate_infections(reported_cases, family = "negbin",
                                    generation_time = generation_time,
                                    incubation_period = incubation_period,
                                    reporting_delay = reporting_delay,
                                    samples = 1000, warmup = 200,
                                    cores = 4, chains = 4,
                                    estimate_rt = TRUE,
                                    verbose = TRUE, return_fit = TRUE)

out
}