Extract Parameter Samples from a Stan Model

Extracts a custom set of parameters from a stan object and adds stratification and dates where appropriate.

Usage

extract_parameter_samples(
  stan_fit,
  data,
  reported_dates,
  imputed_dates,
  reported_inf_dates,
  drop_length_1 = FALSE,
  merge = FALSE
)

Arguments

stan_fit

A <stanfit> or <CmdStanMCMC> object as returned by fit_model().

data

A list of the data supplied to the fit_model() call.

reported_dates

A vector of dates to report estimates for.

imputed_dates

A vector of dates to report imputed reports for.

reported_inf_dates

A vector of dates to report infection estimates for.

drop_length_1

Logical; whether the first dimension should be dropped if it is off length 1; this is necessary when processing simulation results.

merge

if TRUE, merge samples and data so that parameters can be extracted from data.

Value

A list of <data.frame>'s each containing the posterior of a parameter