Perform checks for <dist_spec> get_... functions

Perform checks for <dist_spec> get_... functions

Usage

get_dist_spec_id(x, id)

Arguments

x

A <dist_spec>.

id

Integer; the id of the distribution to get parameters of (if x is a composite distribution). If x is a single distribution this is ignored and can be left as NULL.

Value

The id to use.

Author

Sebastian Funk