Get Combined Regional Results

Get Combined Regional Results

get_regional_results(
  regional_output,
  results_dir,
  date,
  samples = TRUE,
  forecast = FALSE
)

Arguments

regional_output
A list of output as produced by regional_epinow and stored in the regional list.

results_dir
A character string indicating the folder containing the EpiNow2 results to extract.

date
A Character string (in the format "yyyy-mm-dd") indicating the date to extract data for. Defaults to "latest" which finds the latest results available.

samples
Logical, defaults to TRUE. Should samples be returned.

forecast
Logical, defaults to FALSE. Should forecast results be returned.

Value

A list of estimates, forecasts and estimated cases by date of report.

Examples

# \donttest{
# Construct example distributions
generation_time <- get_generation_time(disease = "SARS-CoV-2", source = "ganyani")
incubation_period <- get_incubation_period(disease = "SARS-CoV-2", source = "lauer")
reporting_delay <- EpiNow2::bootstrapped_dist_fit(rlnorm(100, log(6), 1), max_value = 30)

                        
# Uses example case vector from EpiSoon
cases <- EpiNow2::example_confirmed[1:30]

cases <- data.table::rbindlist(list(
  data.table::copy(cases)[, region := "testland"],
  cases[, region := "realland"]))
  
# Run basic nowcasting pipeline
regional_out <- regional_epinow(reported_cases = cases,
                                generation_time = generation_time,
                                delays = list(incubation_period, reporting_delay),
                                stan_args = list(cores = ifelse(interactive(), 4, 1)),
                                summary = FALSE)

#> INFO [2020-10-01 18:58:42] Reporting estimates using data up to: 2020-03-22
#> INFO [2020-10-01 18:58:42] Producing estimates for: testland, realland
#> INFO [2020-10-01 18:58:42] Regions excluded: none
#> INFO [2020-10-01 18:58:42] Initialising estimates for: testland
#> WARN [2020-10-01 19:03:33] testland: There were 3 divergent transitions after warmup. See
#> http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup
#> to find out why this is a problem and how to eliminate them. - 
#> WARN [2020-10-01 19:03:33] testland: Examine the pairs() plot to diagnose sampling problems
#>  - 
#> WARN [2020-10-01 19:03:33] testland: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess - 
#> INFO [2020-10-01 19:03:34] Completed estimates for: testland
#> INFO [2020-10-01 19:03:34] Initialising estimates for: realland
#> WARN [2020-10-01 19:07:49] realland: There were 3 divergent transitions after warmup. See
#> http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup
#> to find out why this is a problem and how to eliminate them. - 
#> WARN [2020-10-01 19:07:49] realland: Examine the pairs() plot to diagnose sampling problems
#>  - 
#> WARN [2020-10-01 19:07:50] realland: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess - 
#> WARN [2020-10-01 19:07:50] realland: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess - 
#> INFO [2020-10-01 19:07:51] Completed estimates for: realland

summary_only <- get_regional_results(regional_out$regional, forecast = FALSE, samples = FALSE)
names(summary_only)

#> [1] "estimates"

all <- get_regional_results(regional_out$regional, forecast = TRUE)
names(all)

#> [1] "estimates"                "forecast"                
#> [3] "estimated_reported_cases"

# }