Summarise Regional Runtimes

Used internally by regional_epinow to summarise region run times.

Usage

regional_runtimes(
  regional_output = NULL,
  target_folder = NULL,
  target_date = NULL,
  return_output = FALSE
)

Arguments

regional_output

A list of output as produced by regional_epinow and stored in the regional list.

target_folder

Character string specifying where to save results (will create if not present).

target_date

A character string giving the target date for which to extract results (in the format "yyyy-mm-dd"). Defaults to latest available estimates.

return_output

Logical, defaults to FALSE. Should output be returned, this automatically updates to TRUE if no directory for saving is specified.

Value

A data.table of region run times

See also

Examples

# \donttest{
# example delays
generation_time <- get_generation_time(
 disease = "SARS-CoV-2", source = "ganyani"
)
incubation_period <- get_incubation_period(
 disease = "SARS-CoV-2", source = "lauer"
)
reporting_delay <- estimate_delay(rlnorm(100, log(6), 1), max_value = 15)
#> Warning: `samples` must be at least 1000. Now setting it to 1000 internally.
#> Warning: `samples` must be at least 1000. Now setting it to 1000 internally.
#> Warning: `samples` must be at least 1000. Now setting it to 1000 internally.
#> Warning: `samples` must be at least 1000. Now setting it to 1000 internally.
#> Warning: `samples` must be at least 1000. Now setting it to 1000 internally.
#> Warning: `samples` must be at least 1000. Now setting it to 1000 internally.
#> Warning: `samples` must be at least 1000. Now setting it to 1000 internally.
#> Warning: `samples` must be at least 1000. Now setting it to 1000 internally.
#> Warning: `samples` must be at least 1000. Now setting it to 1000 internally.
#> Warning: `samples` must be at least 1000. Now setting it to 1000 internally.

cases <- example_confirmed[1:30]
cases <- data.table::rbindlist(list(
  data.table::copy(cases)[, region := "testland"],
  cases[, region := "realland"]
))

# run basic nowcasting pipeline
regional_out <- regional_epinow(
  reported_cases = cases,
  generation_time = generation_time_opts(generation_time),
  delays = delay_opts(incubation_period + reporting_delay),
  stan = stan_opts(samples = 100, warmup = 100),
  output = c("region", "timing")
)
#> INFO [2023-09-26 16:06:51] Producing following optional outputs: regions, timing
#> Logging threshold set at INFO for the EpiNow2 logger
#> Writing EpiNow2 logs to the console and: /tmp/RtmpH53zkW/regional-epinow/2020-03-22.log
#> Logging threshold set at INFO for the EpiNow2.epinow logger
#> Writing EpiNow2.epinow logs to: /tmp/RtmpH53zkW/epinow/2020-03-22.log
#> INFO [2023-09-26 16:06:51] Reporting estimates using data up to: 2020-03-22
#> INFO [2023-09-26 16:06:51] No target directory specified so returning output
#> INFO [2023-09-26 16:06:51] Producing estimates for: testland, realland
#> INFO [2023-09-26 16:06:51] Regions excluded: none
#> INFO [2023-09-26 16:07:21] Completed estimates for: testland
#> INFO [2023-09-26 16:07:56] Completed estimates for: realland
#> INFO [2023-09-26 16:07:56] Completed regional estimates
#> INFO [2023-09-26 16:07:56] Regions with estimates: 2
#> INFO [2023-09-26 16:07:56] Regions with runtime errors: 0
#> INFO [2023-09-26 16:07:56] No target directory specified so returning timings

regional_runtimes(regional_output = regional_out$regional)
#> INFO [2023-09-26 16:07:56] No target directory specified so returning timings
#>      region time
#> 1: testland 30.7
#> 2: realland 34.5
# }