Run epinow with Regional Processing Code

Internal function that handles calling epinow. Future work will extend this function to better handle stan logs and allow the user to modify settings between regions.

run_region(
  target_region,
  generation_time,
  delays,
  truncation,
  rt,
  backcalc,
  gp,
  obs,
  stan,
  horizon,
  CrIs,
  reported_cases,
  target_folder,
  target_date,
  return_output,
  output,
  complete_logger,
  verbose,
  progress_fn,
  ...
)

Arguments

target_region
Character string indicating the region being evaluated

generation_time
A list containing the mean, standard deviation of the mean (mean_sd), standard deviation (sd), standard deviation of the standard deviation and the maximum allowed value for the generation time (assuming a gamma distribution).

delays
A call to delay_opts() defining delay distributions and options. See the documentation of delay_opts() and the examples below for details.

truncation
A list of options as generated by trunc_opts() defining the truncation of observed data. Defaults to trunc_opts(). See estimate_truncation() for an approach to estimating truncation from data.

rt
A list of options as generated by rt_opts() defining Rt estimation. Defaults to rt_opts(). Set to NULL to switch to using back calculation rather than generating infections using Rt.

backcalc
A list of options as generated by backcalc_opts() to define the back calculation. Defaults to backcalc_opts().

gp
A list of options as generated by gp_opts() to define the Gaussian process. Defaults to gp_opts().Set to NULL to disable the Gaussian process.

obs
A list of options as generated by obs_opts() defining the observation model. Defaults to obs_opts().

stan
A list of stan options as generated by stan_opts(). Defaults to stan_opts(). Can be used to override data, init, and verbose settings if desired.

horizon
Numeric, defaults to 7. Number of days into the future to forecast.

CrIs
Numeric vector of credible intervals to calculate.

reported_cases
A data frame of confirmed cases (confirm) by date (date), and region (region).

target_folder
Character string specifying where to save results (will create if not present).

target_date
Date, defaults to maximum found in the data if not specified.

return_output
Logical, defaults to FALSE. Should output be returned, this automatically updates to TRUE if no directory for saving is specified.

output
A character vector of optional output to return. Supported options are the individual regional estimates ("regions"), samples ("samples"), plots ("plots"), copying the individual region dated folder into a latest folder (if target_folder is not null, set using "latest"), the stan fit of the underlying model ("fit"), and an overall summary across regions ("summary"). The default is to return samples and plots alongside summarised estimates and summary statistics. If target_folder is not NULL then the default is also to copy all results into a latest folder.

complete_logger
Character string indicating the logger to output the completion of estimation to.

verbose
Logical defaults to FALSE. Outputs verbose progress messages to the console from epinow.

progress_fn
Function as returned by progressr::progressor. Allows the use of a progress bar.

...
Pass additional arguments to epinow. See the documentation for epinow for details.

Value

A list of processed output as produced by process_region

See also