Set up Future Backend

A utility function that aims to streamline the set up of the required future backend with sensible defaults for most users of regional_epinow. More advanced users are recommended to setup their own future backend based on their available resources.

setup_future(
  reported_cases,
  strategies = c("multiprocess", "multiprocess"),
  min_cores_per_worker = 4
)

Arguments

reported_cases
A data frame of confirmed cases (confirm) by date (date), and region (region).

strategies
A vector length 1 to 2 of strategies to pass to future::plan. Nesting of parallisation is from the top level down. The default is to set up nesting parallisation with both using future::multiprocess. For single level parallisation use a single strategy or future::plan directly. See ?future::plan for options.

min_cores_per_worker
Numeric, the minimum number of cores per worker. Defaults to 4 which assumes 4 MCMC chains are in use per region.

Value

Numeric number of cores to use per worker. If greater than 1 pass to stan_args = list(cores = "output from setup future") or use future = TRUE. If only a single strategy is used then nothing is returned.