Fit a Subsampled Boostrap to Integer Values and Summarise Distribution Parameters

Fit a Subsampled Boostrap to Integer Values and Summarise Distribution Parameters

bootstrapped_dist_fit(
  values,
  dist = "lognormal",
  samples = 2000,
  bootstraps = 10,
  bootstrap_samples = 250,
  verbose = FALSE
)

Arguments

values
Numeric vector of integer values.

dist
Character string, which distribution to fit. Defaults to lognormal ("lognormal") but gamma ("gamma") is also supported.

samples
Numeric, number of samples to take overall from the bootstrapped posteriors.

bootstraps
Numeric, defaults to 1. The number of bootstrap samples (with replacement) of the delay distribution to take.

bootstrap_samples
Numeric, defaults to 100. The number of samples to take in each boostrap. When the sample size of the supplied delay distribution is less than 100 this is used instead.

verbose
Logical, defaults to FALSE. Should progress messages be printed

Value

A list summarising the bootstrapped distribution

Examples

if (FALSE) {
# lognormal
delays <- rlnorm(500, log(5), 1)

out <- bootstrapped_dist_fit(delays, samples = 1000, bootstraps = 10,
                             dist = "lognormal")

## Inspect
out

}