Create a List of Stan Arguments

Generates a list of arguments as required by rstan::sampling (when method = "exact) or rstan::vb (when method = "approximate). See create_stan_args() for the defaults and the relevant rstan functions for additional options.

create_stan_args(
  model,
  data = NULL,
  init = "random",
  samples = 1000,
  stan_args = NULL,
  method = "exact",
  verbose = FALSE
)

Arguments

model
A stan model object, defaults to packaged model if not supplied.

data
A list of stan data as created by create_stan_data

init
Initial conditions passed to rstan. Defaults to "random" but can also be a function ( as supplied by create_intitial_conditions).

samples
Numeric, defaults to 1000. The overall number of posterior samples to return (Note: not the number of samples per chain as is the default in stan).

stan_args
A list of stan arguments to be passed to rstan::sampling or rstan::vb (when using the "exact" or "approximate" method). For method = approximate an additional argument trials indicates the number of attempts to make using variational inference before returning an error (as stochastic failure is possible). The default for this is 5.

method
A character string defaults to "exact". Also accepts "approximate". Indicates the fitting method to be used this can either be "exact" (NUTs sampling) or "approximate" (variational inference). The exact approach returns samples from the posterior whilst the approximate method returns approximate samples. The approximate method is likely to return results several order of magnitudes faster than the exact method.

verbose
Logical, defaults to FALSE. Should verbose progress messages be returned.

Value

A list of stan arguments

Examples

# default settings
create_stan_args()

#> $object
#> S4 class stanmodel 'estimate_infections' coded as follows:
#> functions {
#> 
#>  // discretised truncated gamma pmf
#>   real discretised_gamma_pmf(int y, real mu, real sigma, int max_val) {
#>     // calculate alpha and beta for gamma distribution
#>     real c_sigma = sigma + 1e-5;
#>     real alpha = ((mu)/ c_sigma)^2;
#>     real beta = (mu) / (c_sigma^2);
#>     //account for numerical issues
#>     alpha = alpha <= 0 ? 1e-5 : alpha;
#>     beta = beta <= 0 ? 1e-5 : beta;
#>     alpha = is_inf(alpha) ? 1e8 : alpha;
#>     beta = is_inf(beta) ? 1e8 : beta; 
#>     return((gamma_cdf(y + 1, alpha, beta) - gamma_cdf(y, alpha, beta)) / 
#>     (gamma_cdf(max_val + 1, alpha, beta) - gamma_cdf(1, alpha, beta)));
#>   }
#> 
#> 
#>   // discretised truncated lognormal pmf
#>   real discretised_lognormal_pmf(int y, real mu, real sigma, int max_val) {
#>     real small = 1e-5;
#>     real adj_y = y + small;
#>     return((normal_cdf((log(adj_y + 1) - mu) / sigma, 0.0, 1.0) -
#>             normal_cdf((log(adj_y) - mu) / sigma, 0.0, 1.0)) /
#>            (normal_cdf((log(max_val + small) - mu) / sigma, 0.0, 1.0) -
#>             normal_cdf((log(small) - mu) / sigma, 0.0, 1.0)));
#>   }
#> 
#> // convolve a pdf and case vector 
#> vector convolve(vector cases, vector pdf) {
#>     int t = num_elements(cases);
#>     int max_pdf = num_elements(pdf);
#>     vector[t] convolved_cases = rep_vector(1e-5, t);
#>     for (s in 1:t) {
#>         convolved_cases[s] += dot_product(cases[max(1, (s - max_pdf + 1)):s], tail(pdf, min(max_pdf, s)));
#>     }
#>    return(convolved_cases);
#>   }
#> 
#> // convolve latent infections to reported (but still unobserved) cases
#> vector convolve_to_report(vector infections, 
#>                           real[] delay_mean, 
#>                           real[] delay_sd,
#>                           int[] max_delay,
#>                           int seeding_time) {
#>   int t = num_elements(infections);
#>   vector[t - seeding_time] reports;
#>   vector[t] reports_hold = infections;
#>   int delays = num_elements(delay_mean);
#>   if (delays) {
#>     for (s in 1:delays) {
#>       vector[max_delay[s]] rev_delay = rep_vector(1e-5, max_delay[s]);
#>       for (j in 1:(max_delay[s])) {
#>         rev_delay[j] +=
#>         discretised_lognormal_pmf(max_delay[s] - j, delay_mean[s], delay_sd[s], max_delay[s]);
#>       }
#>       reports_hold = convolve(reports_hold, rev_delay);
#>     }
#>     reports = reports_hold[(seeding_time + 1):t];
#>   }else{
#>     reports = infections[(seeding_time + 1):t];
#>   }
#>   return(reports);
#> }
#> 
#> void delays_lp(real[] delay_mean, real[] delay_mean_mean, real[] delay_mean_sd, 
#>                real[] delay_sd, real[] delay_sd_mean, real[] delay_sd_sd, int weight){
#>     int delays = num_elements(delay_mean);
#>     if (delays) {
#>     for (s in 1:delays) {
#>       target += normal_lpdf(delay_mean[s] | delay_mean_mean[s], delay_mean_sd[s]) * weight;
#>       target += normal_lpdf(delay_sd[s] | delay_sd_mean[s], delay_sd_sd[s]) * weight;
#>     }
#>   }
#> }
#> 
#>   // exponential quadratic kernal
#> 	real spd_SE(real alpha, real rho, real w) {
#> 		real S;
#> 		S = (alpha^2) * sqrt(2*pi()) * rho * exp(-0.5*(rho^2)*(w^2));
#> 		return S;
#> 	}
#> 	
#> 	// basis function for approximate hilbert space gp
#> 	// see here for details: https://arxiv.org/pdf/2004.11408.pdf
#> 	vector phi_SE(real L, int m, vector x) {
#> 		vector[rows(x)] fi;
#> 		fi = 1/sqrt(L) * sin(m*pi()/(2*L) * (x+L));
#> 		return fi;
#> 	}
#> 	
#> 	// eigenvalues for approximate hilbert space gp
#> 	// see here for details: https://arxiv.org/pdf/2004.11408.pdf
#> 	real lambda(real L, int m) {
#> 		real lam;
#> 		lam = ((m*pi())/(2*L))^2;
#> 		return lam;
#> 	}
#> 
#> 
#> int setup_noise(int ot_h, int t, int horizon, int estimate_r, 
#>                 int stationary, int future_fixed, int fixed_from) {
#>   int noise_time = estimate_r > 0 ? (stationary > 0 ? ot_h : ot_h - 1) : t;
#>   int noise_terms =  future_fixed > 0 ? (noise_time - horizon + fixed_from) : noise_time; 
#>   return(noise_terms);
#> }
#> 
#> 
#> matrix setup_gp(int M, real L, int dimension) {
#>   vector[dimension] time;
#>   matrix[dimension, M] PHI;
#>   for (s in 1:dimension) {
#>     time[s] = s;
#>   }
#>   for (m in 1:M){ 
#>     PHI[,m] = phi_SE(L, m, time); 
#>   }
#>   return(PHI);
#> }
#> 
#> 
#> vector update_gp(matrix PHI, int M, real L, real alpha, real rho, vector eta) {
#>   vector[M] diagSPD;    // spectral density
#>   vector[M] SPD_eta;    // spectral density * noise
#>   int noise_terms = rows(PHI);
#>   vector[noise_terms] noise = rep_vector(1e-6, noise_terms);
#>   // GP in noise - spectral densities
#>   for(m in 1:M){ 
#>     diagSPD[m] =  sqrt(spd_SE(alpha, rho, sqrt(lambda(L, m)))); 
#>   }
#>   SPD_eta = diagSPD .* eta;
#>   noise = noise + PHI[,] * SPD_eta;
#>   return(noise);
#> }
#> 
#> void gaussian_process_lp(real[] rho, real[] alpha, vector eta,
#>                          real ls_alpha, real ls_beta, real alpha_sd) {
#>   rho ~ inv_gamma(ls_alpha, ls_beta);
#>   alpha ~ normal(0, alpha_sd);
#>   eta ~ std_normal();
#> }
#> 
#> 
#> 
#> real update_breakpoints(real input_R, real[] bp_effects,
#>                         int bp_index, int at_bp,
#>                         int stationary) {
#>   real R = input_R;
#>   if (stationary) {
#>     if (bp_index > 0) {
#>       R += sum(bp_effects[1:bp_index]);
#>       }
#>   }else{
#>     if (at_bp) {
#>       R += bp_effects[bp_index];
#>     }
#>   }               
#>   return(R);
#> }
#> 
#> real update_R(vector R, vector noise, int noise_terms,
#>               int index, int stationary) {
#>   real cR = R[index];
#>   if (noise_terms > 0) {
#>     if (stationary){
#>       if (index <= noise_terms) {
#>         cR += noise[index];
#>       }else{
#>         if (index > 1) {
#>            cR = R[index - 1];
#>         }
#>       }
#>     }else{
#>       if (index <= (noise_terms + 1)) {
#>         cR = R[index - 1] + noise[index - 1];
#>       }else{
#>         cR = R[index - 1];
#>       }
#>     }
#>   }
#>   return(cR);
#> }
#> 
#> vector update_Rt(vector input_R, real log_R, vector noise, int[] bps,
#>                  real[] bp_effects, int stationary) {
#>   // define control parameters
#>   int noise_terms = num_elements(noise);
#>   int i_stationary = noise_terms > 0 ? stationary : 1;
#>   int t = num_elements(input_R);
#>   int bp_n = num_elements(bp_effects);
#>   int bp_in = 0;
#>   int at_bp = 0;
#>   int index;
#>   vector[t] R;
#>   // initialise Rt
#>   if (i_stationary) {
#>     R = rep_vector(log_R, t);
#>     index = 1;
#>   }else{
#>     R[1] = log_R;
#>     index = 2;
#>   }
#>   // iteratively update Rt
#>   for (s in index:t) {
#>     R[s] = update_R(R, noise, noise_terms, s, i_stationary);
#>     if (bp_n > 0) {
#>       at_bp = bps[s];
#>       bp_in += at_bp;
#>       R[s] = update_breakpoints(R[s], bp_effects, bp_in, at_bp, i_stationary);
#>     }
#>   }
#>   // convert to correct scale
#>   R = exp(R);
#>   return(R);
#> }
#> 
#> 
#> real update_infectiousness(vector infections, vector gt_pmf,
#>                            int seeding_time, int max_gt, int index){
#>   int inf_start = max(1, (index + seeding_time - max_gt));
#>   int inf_end = (index + seeding_time - 1);
#>   int pmf_accessed = min(max_gt, index + seeding_time - 1);
#>   real new_inf = dot_product(infections[inf_start:inf_end], tail(gt_pmf, pmf_accessed));
#>   return(new_inf);
#> }
#> 
#> vector generate_infections(vector R, int seeding_time, real[] gt_mean, 
#>                            real[] gt_sd, int max_gt, vector shifted_cases,
#>                            vector initial_infections) {
#>   int rt = num_elements(R);
#>   int t = rt + seeding_time;
#>   vector[t] infections = rep_vector(1e-5, t);
#>   vector[max_gt] gt_pmf;              // reversed generation time pdf
#>   vector[rt] infectiousness = rep_vector(1e-5, rt);  // infections over time
#>   
#>   for (j in 1:(max_gt)) {
#>     gt_pmf[j] = discretised_gamma_pmf(max_gt - j + 1, gt_mean[1], gt_sd[1], max_gt);
#>   }
#>   // estimate initial infections not using Rt
#>   infections[1:seeding_time] = infections[1:seeding_time] + shifted_cases[1:seeding_time] .* initial_infections;
#>   
#>   for (s in 1:rt) {
#>     infectiousness[s] += update_infectiousness(infections, gt_pmf, seeding_time, max_gt, s);
#>     infections[s + seeding_time] += R[s] * infectiousness[s];
#>   }
#>   return(infections);
#> }
#> 
#> 
#> vector deconvolve_infections(vector shifted_cases, vector noise, int fixed) {
#>   int t = num_elements(shifted_cases);
#>   vector[t] infections = rep_vector(1e-5, t);
#>   if(!fixed) {
#>     infections = infections + shifted_cases .* exp(noise);
#>   }else{
#>     infections = infections + shifted_cases;
#>   }
#>   return(infections);
#> }
#> 
#> void generation_time_lp(real[] gt_mean, real gt_mean_mean, real gt_mean_sd, 
#>                         real[] gt_sd, real gt_sd_mean, real gt_sd_sd, int weight) {
#>     target += normal_lpdf(gt_mean[1] | gt_mean_mean, gt_mean_sd) * weight;
#>     target += normal_lpdf(gt_sd[1] | gt_sd_mean, gt_sd_sd) * weight;
#> }
#> 
#> vector calculate_Rt(vector infections, int seeding_time,
#>                     real gt_mean, real gt_sd, int max_gt) {
#>   vector[max_gt] gt_pmf;  
#>   int t = num_elements(infections);
#>   int ot = t - seeding_time;
#>   vector[ot] R;
#>   vector[ot] infectiousness = rep_vector(1e-5, ot); 
#>   // calculate PMF of the generation time
#>   for (j in 1:(max_gt)) {
#>     gt_pmf[j] =  discretised_gamma_pmf(max_gt - j + 1, gt_mean, gt_sd, max_gt);
#>   }
#>   // calculate Rt using Cori et al. approach
#>   for (s in 1:ot) {
#>     infectiousness[s] += update_infectiousness(infections, gt_pmf, seeding_time, max_gt, s);
#>     R[s] = infections[s + seeding_time] / infectiousness[s];
#>   }
#>   return(R);
#> }
#> 
#> 
#> vector day_of_week_effect(vector reports, int[] day_of_week, vector effect) {
#>   int t = num_elements(reports);
#>   // scale day of week effect
#>   vector[7] scaled_effect = 7 * effect;
#>   vector[t] scaled_reports;
#>   for (s in 1:t) {
#>     // add reporting effects (adjust for simplex scale)
#>     scaled_reports[s] = reports[s] * scaled_effect[day_of_week[s]];
#>    }
#>   return(scaled_reports);
#> }
#> 
#>    
#> void report_lp(int[] cases, vector reports, 
#>                real[] rep_phi, int phi_prior,
#>                int model_type, int horizon,
#>                int weight) {
#>   int t = num_elements(reports) - horizon;
#>   if (model_type) {
#>     //overdispersion
#>     rep_phi[model_type] ~ exponential(phi_prior);
#>     target += neg_binomial_2_lpmf(cases | reports[1:t], rep_phi[model_type]) * weight;
#>   }else{
#>     target += poisson_lpmf(cases | reports[1:t]) * weight;
#>   }
#> }
#> 
#> 
#> real[] R_to_growth(vector R, real gt_mean, real gt_sd) {
#>   real k = pow(gt_sd / gt_mean, 2);
#>   int t = num_elements(R);
#>   real r[t];
#>   for (s in 1:t) {
#>     r[s] = (pow(R[s], k) - 1) / (k * gt_mean);
#>   } 
#>   return(r);
#> }
#> 
#> int[] report_rng(vector reports, real[] rep_phi, int model_type) {
#>   int t = num_elements(reports);
#>   int sampled_reports[t];
#>   if (model_type) {
#>     for (s in 1:t) {
#>       sampled_reports[s] = neg_binomial_2_rng(reports[s] > 1e8 ? 1e8 : reports[s], rep_phi[model_type]);
#>     }
#>   }else{
#>     for (s in 1:t) {
#>       sampled_reports[s] = poisson_rng(reports[s] > 1e8 ? 1e8 : reports[s]);
#>     }
#>   }
#>   return(sampled_reports);
#> }
#> 
#> 
#> }
#> 
#> 
#> data {
#> 
#>   int t;                                            // unobserved time
#>   int seeding_time;                                 // time period used for seeding and not observed
#>   int horizon;                                      // forecast horizon
#>   int<lower = 0> cases[t - horizon - seeding_time]; // observed cases
#>   vector<lower = 0>[t] shifted_cases;               // median shifted smoothed cases
#> 
#>   int delays;                  // no. of delay distributions
#>   real delay_mean_sd[delays];  // prior sd of mean incubation period
#>   real delay_mean_mean[delays];// prior mean of mean incubation period
#>   real delay_sd_mean[delays];  // prior sd of sd of incubation period
#>   real delay_sd_sd[delays];    // prior sd of sd of incubation period
#>   int max_delay[delays];       // maximum incubation period
#> 
#>   real L;				                     // boundary value for infections gp
#>   int<lower=1> M;			               // basis functions for infections gp
#>   real lengthscale_alpha;            // alpha for gp lengthscale prior
#>   real lengthscale_beta;             // beta for gp lengthscale prior
#>   real alpha_sd;                     // standard deviation of the alpha gp kernal parameter
#>   int stationary;                    // is underlying gaussian process first or second order
#>   int fixed;                         //  should a gaussian process be used
#> 
#>   real gt_mean_sd;                   // prior sd of mean generation time
#>   real gt_mean_mean;                 // prior mean of mean generation time
#>   real gt_sd_mean;                   // prior sd of sd of generation time
#>   real gt_sd_sd;                     // prior sd of sd of generation time
#>   int max_gt;                        // maximum generation time
#> 
#>   int estimate_r;                    // should the reproduction no be estimated (1 = yes)
#>   real <lower = 0> r_mean;           // prior mean of reproduction number
#>   real <lower = 0> r_sd;             // prior standard deviation of reproduction number
#>   int bp_n;                          // no of breakpoints (0 = no breakpoints)
#>   int breakpoints[t - seeding_time]; // when do breakpoints occur 
#>   int future_fixed;                  // is underlying future Rt assumed to be fixed
#>   int fixed_from;                    // Reference date for when Rt estimation should be fixed
#> 
#>   int day_of_week[t - seeding_time]; // day of the week indicator (1 - 7)
#>   int model_type;                    // type of model: 0 = poisson otherwise negative binomial
#>   int week_effect;                   // should a day of the week effect be estimated
#> }
#> 
#> transformed data{
#>   // observations
#>   int ot = t - seeding_time - horizon;  // observed time
#>   int ot_h = ot + horizon;  // observed time + forecast horizon
#>   // gaussian process
#>   int noise_terms = setup_noise(ot_h, t, horizon, estimate_r, stationary, future_fixed, fixed_from);
#>   matrix[noise_terms, M] PHI = setup_gp(M, L, noise_terms);  // basis function 
#>   // Rt
#>   real r_logmean = log(r_mean^2 / sqrt(r_sd^2 + r_mean^2)); 
#>   real r_logsd = sqrt(log(1 + (r_sd^2 / r_mean^2))); 
#> }
#> 
#> parameters{
#>   // gaussian process
#>   real<lower = 0> rho[fixed ? 0 : 1];    // length scale of noise GP
#>   real<lower = 0> alpha[fixed ? 0 : 1];  // scale of of noise GP
#>   vector[fixed ? 0 : M] eta;             // unconstrained noise
#>   // Rt
#>   vector[estimate_r] log_R;             // baseline reproduction number estimate (log)
#>   vector[estimate_r > 0 ? seeding_time : 0] initial_infections;// seed infections 
#>   real<lower = 0> gt_mean[estimate_r];  // mean of generation time
#>   real <lower = 0> gt_sd[estimate_r];   // sd of generation time
#>   real bp_effects[bp_n];                // Rt breakpoint effects
#>   // observation model
#>   real<lower = 0> delay_mean[delays];   // mean of delays
#>   real<lower = 0> delay_sd[delays];     // sd of delays
#>   simplex[week_effect ? 7 : 1] day_of_week_simplex;   // day of week reporting effect 
#>   real<lower = 0> rep_phi[model_type];  // overdispersion of the reporting process
#> }
#> 
#> transformed parameters {
#>   vector[fixed ? 0 : noise_terms] noise;                  // noise  generated by the gaussian process
#>   vector[estimate_r > 0 ? ot_h : 0] R;                    // reproduction number
#>   vector[t] infections;                                   // latent infections
#>   vector[ot_h] reports;                                   // observed cases
#>   // GP in noise - spectral densities
#>   if (!fixed) {
#>     noise = update_gp(PHI, M, L, alpha[1], rho[1], eta);
#>   }
#>   // Estimate latent infections
#>   if (estimate_r) {
#>     // via Rt
#>     R = update_Rt(R, log_R[estimate_r], noise, breakpoints, bp_effects, stationary);
#>     infections = generate_infections(R, seeding_time, gt_mean, gt_sd, max_gt, shifted_cases, initial_infections);
#>   }else{
#>     // via deconvolution
#>     infections = deconvolve_infections(shifted_cases, noise, fixed);
#>   }
#>   // convolve from latent infections to mean of observations
#>   reports = convolve_to_report(infections, delay_mean, delay_sd, max_delay, seeding_time);
#>  // weekly reporting effect
#>  if (week_effect) {
#>    reports = day_of_week_effect(reports, day_of_week, day_of_week_simplex);
#>   }
#> }
#> 
#> model {
#>   // priors for noise GP
#>   if (!fixed) {
#>     gaussian_process_lp(rho, alpha, eta, lengthscale_alpha, lengthscale_beta, alpha_sd);
#>   }
#>   // penalised priors for delay distributions
#>   delays_lp(delay_mean, delay_mean_mean, delay_mean_sd, delay_sd, delay_sd_mean, delay_sd_sd, t);
#>   // Rt priors
#>   if (estimate_r) {
#>     // prior on R
#>     log_R ~ normal(r_logmean, r_logsd);
#>     //breakpoint effects on Rt
#>     if (bp_n > 0) {
#>       bp_effects ~ normal(0, 0.1);
#>     }
#>     // initial infections
#>     initial_infections ~ lognormal(0, 0.1);
#>     // penalised_prior on generation interval
#>     generation_time_lp(gt_mean, gt_mean_mean, gt_mean_sd, gt_sd, gt_sd_mean, gt_sd_sd, ot);
#>   }
#>   // observed reports from mean of reports
#>   report_lp(cases, reports, rep_phi, 1, model_type, horizon, 1);
#> }
#>   
#> generated quantities {
#>   int imputed_reports[ot_h]; 
#>   vector[estimate_r > 0 ? 0: ot_h] gen_R;
#>   real r[ot_h];
#>   if (estimate_r){
#>     // estimate growth from estimated Rt
#>     r = R_to_growth(R, gt_mean[1], gt_sd[1]);
#>   }else{
#>     // sample generation time
#>     real gt_mean_sample = normal_rng(gt_mean_mean, gt_mean_sd);
#>     real gt_sd_sample = normal_rng(gt_sd_mean, gt_sd_sd);
#>     // calculate Rt using infections and generation time
#>     gen_R = calculate_Rt(infections, seeding_time, gt_mean_sample, gt_mean_sample, max_gt);
#>     // estimate growth from calculated Rt
#>     r = R_to_growth(gen_R, gt_mean_sample, gt_sd_sample);
#>   }
#>   // simulate reported cases
#>   imputed_reports = report_rng(reports, rep_phi, model_type);
#> } 
#> 
#> $data
#> NULL
#> 
#> $init
#> [1] "random"
#> 
#> $refresh
#> [1] 0
#> 
#> $cores
#> [1] 4
#> 
#> $warmup
#> [1] 500
#> 
#> $chains
#> [1] 4
#> 
#> $control
#> $control$adapt_delta
#> [1] 0.99
#> 
#> $control$max_treedepth
#> [1] 15
#> 
#> 
#> $save_warmup
#> [1] FALSE
#> 
#> $seed
#> [1] 67020432
#> 
#> $iter
#> [1] 750
#> 

# approximate settings
create_stan_args(method = "approximate") 

#> $object
#> S4 class stanmodel 'estimate_infections' coded as follows:
#> functions {
#> 
#>  // discretised truncated gamma pmf
#>   real discretised_gamma_pmf(int y, real mu, real sigma, int max_val) {
#>     // calculate alpha and beta for gamma distribution
#>     real c_sigma = sigma + 1e-5;
#>     real alpha = ((mu)/ c_sigma)^2;
#>     real beta = (mu) / (c_sigma^2);
#>     //account for numerical issues
#>     alpha = alpha <= 0 ? 1e-5 : alpha;
#>     beta = beta <= 0 ? 1e-5 : beta;
#>     alpha = is_inf(alpha) ? 1e8 : alpha;
#>     beta = is_inf(beta) ? 1e8 : beta; 
#>     return((gamma_cdf(y + 1, alpha, beta) - gamma_cdf(y, alpha, beta)) / 
#>     (gamma_cdf(max_val + 1, alpha, beta) - gamma_cdf(1, alpha, beta)));
#>   }
#> 
#> 
#>   // discretised truncated lognormal pmf
#>   real discretised_lognormal_pmf(int y, real mu, real sigma, int max_val) {
#>     real small = 1e-5;
#>     real adj_y = y + small;
#>     return((normal_cdf((log(adj_y + 1) - mu) / sigma, 0.0, 1.0) -
#>             normal_cdf((log(adj_y) - mu) / sigma, 0.0, 1.0)) /
#>            (normal_cdf((log(max_val + small) - mu) / sigma, 0.0, 1.0) -
#>             normal_cdf((log(small) - mu) / sigma, 0.0, 1.0)));
#>   }
#> 
#> // convolve a pdf and case vector 
#> vector convolve(vector cases, vector pdf) {
#>     int t = num_elements(cases);
#>     int max_pdf = num_elements(pdf);
#>     vector[t] convolved_cases = rep_vector(1e-5, t);
#>     for (s in 1:t) {
#>         convolved_cases[s] += dot_product(cases[max(1, (s - max_pdf + 1)):s], tail(pdf, min(max_pdf, s)));
#>     }
#>    return(convolved_cases);
#>   }
#> 
#> // convolve latent infections to reported (but still unobserved) cases
#> vector convolve_to_report(vector infections, 
#>                           real[] delay_mean, 
#>                           real[] delay_sd,
#>                           int[] max_delay,
#>                           int seeding_time) {
#>   int t = num_elements(infections);
#>   vector[t - seeding_time] reports;
#>   vector[t] reports_hold = infections;
#>   int delays = num_elements(delay_mean);
#>   if (delays) {
#>     for (s in 1:delays) {
#>       vector[max_delay[s]] rev_delay = rep_vector(1e-5, max_delay[s]);
#>       for (j in 1:(max_delay[s])) {
#>         rev_delay[j] +=
#>         discretised_lognormal_pmf(max_delay[s] - j, delay_mean[s], delay_sd[s], max_delay[s]);
#>       }
#>       reports_hold = convolve(reports_hold, rev_delay);
#>     }
#>     reports = reports_hold[(seeding_time + 1):t];
#>   }else{
#>     reports = infections[(seeding_time + 1):t];
#>   }
#>   return(reports);
#> }
#> 
#> void delays_lp(real[] delay_mean, real[] delay_mean_mean, real[] delay_mean_sd, 
#>                real[] delay_sd, real[] delay_sd_mean, real[] delay_sd_sd, int weight){
#>     int delays = num_elements(delay_mean);
#>     if (delays) {
#>     for (s in 1:delays) {
#>       target += normal_lpdf(delay_mean[s] | delay_mean_mean[s], delay_mean_sd[s]) * weight;
#>       target += normal_lpdf(delay_sd[s] | delay_sd_mean[s], delay_sd_sd[s]) * weight;
#>     }
#>   }
#> }
#> 
#>   // exponential quadratic kernal
#> 	real spd_SE(real alpha, real rho, real w) {
#> 		real S;
#> 		S = (alpha^2) * sqrt(2*pi()) * rho * exp(-0.5*(rho^2)*(w^2));
#> 		return S;
#> 	}
#> 	
#> 	// basis function for approximate hilbert space gp
#> 	// see here for details: https://arxiv.org/pdf/2004.11408.pdf
#> 	vector phi_SE(real L, int m, vector x) {
#> 		vector[rows(x)] fi;
#> 		fi = 1/sqrt(L) * sin(m*pi()/(2*L) * (x+L));
#> 		return fi;
#> 	}
#> 	
#> 	// eigenvalues for approximate hilbert space gp
#> 	// see here for details: https://arxiv.org/pdf/2004.11408.pdf
#> 	real lambda(real L, int m) {
#> 		real lam;
#> 		lam = ((m*pi())/(2*L))^2;
#> 		return lam;
#> 	}
#> 
#> 
#> int setup_noise(int ot_h, int t, int horizon, int estimate_r, 
#>                 int stationary, int future_fixed, int fixed_from) {
#>   int noise_time = estimate_r > 0 ? (stationary > 0 ? ot_h : ot_h - 1) : t;
#>   int noise_terms =  future_fixed > 0 ? (noise_time - horizon + fixed_from) : noise_time; 
#>   return(noise_terms);
#> }
#> 
#> 
#> matrix setup_gp(int M, real L, int dimension) {
#>   vector[dimension] time;
#>   matrix[dimension, M] PHI;
#>   for (s in 1:dimension) {
#>     time[s] = s;
#>   }
#>   for (m in 1:M){ 
#>     PHI[,m] = phi_SE(L, m, time); 
#>   }
#>   return(PHI);
#> }
#> 
#> 
#> vector update_gp(matrix PHI, int M, real L, real alpha, real rho, vector eta) {
#>   vector[M] diagSPD;    // spectral density
#>   vector[M] SPD_eta;    // spectral density * noise
#>   int noise_terms = rows(PHI);
#>   vector[noise_terms] noise = rep_vector(1e-6, noise_terms);
#>   // GP in noise - spectral densities
#>   for(m in 1:M){ 
#>     diagSPD[m] =  sqrt(spd_SE(alpha, rho, sqrt(lambda(L, m)))); 
#>   }
#>   SPD_eta = diagSPD .* eta;
#>   noise = noise + PHI[,] * SPD_eta;
#>   return(noise);
#> }
#> 
#> void gaussian_process_lp(real[] rho, real[] alpha, vector eta,
#>                          real ls_alpha, real ls_beta, real alpha_sd) {
#>   rho ~ inv_gamma(ls_alpha, ls_beta);
#>   alpha ~ normal(0, alpha_sd);
#>   eta ~ std_normal();
#> }
#> 
#> 
#> 
#> real update_breakpoints(real input_R, real[] bp_effects,
#>                         int bp_index, int at_bp,
#>                         int stationary) {
#>   real R = input_R;
#>   if (stationary) {
#>     if (bp_index > 0) {
#>       R += sum(bp_effects[1:bp_index]);
#>       }
#>   }else{
#>     if (at_bp) {
#>       R += bp_effects[bp_index];
#>     }
#>   }               
#>   return(R);
#> }
#> 
#> real update_R(vector R, vector noise, int noise_terms,
#>               int index, int stationary) {
#>   real cR = R[index];
#>   if (noise_terms > 0) {
#>     if (stationary){
#>       if (index <= noise_terms) {
#>         cR += noise[index];
#>       }else{
#>         if (index > 1) {
#>            cR = R[index - 1];
#>         }
#>       }
#>     }else{
#>       if (index <= (noise_terms + 1)) {
#>         cR = R[index - 1] + noise[index - 1];
#>       }else{
#>         cR = R[index - 1];
#>       }
#>     }
#>   }
#>   return(cR);
#> }
#> 
#> vector update_Rt(vector input_R, real log_R, vector noise, int[] bps,
#>                  real[] bp_effects, int stationary) {
#>   // define control parameters
#>   int noise_terms = num_elements(noise);
#>   int i_stationary = noise_terms > 0 ? stationary : 1;
#>   int t = num_elements(input_R);
#>   int bp_n = num_elements(bp_effects);
#>   int bp_in = 0;
#>   int at_bp = 0;
#>   int index;
#>   vector[t] R;
#>   // initialise Rt
#>   if (i_stationary) {
#>     R = rep_vector(log_R, t);
#>     index = 1;
#>   }else{
#>     R[1] = log_R;
#>     index = 2;
#>   }
#>   // iteratively update Rt
#>   for (s in index:t) {
#>     R[s] = update_R(R, noise, noise_terms, s, i_stationary);
#>     if (bp_n > 0) {
#>       at_bp = bps[s];
#>       bp_in += at_bp;
#>       R[s] = update_breakpoints(R[s], bp_effects, bp_in, at_bp, i_stationary);
#>     }
#>   }
#>   // convert to correct scale
#>   R = exp(R);
#>   return(R);
#> }
#> 
#> 
#> real update_infectiousness(vector infections, vector gt_pmf,
#>                            int seeding_time, int max_gt, int index){
#>   int inf_start = max(1, (index + seeding_time - max_gt));
#>   int inf_end = (index + seeding_time - 1);
#>   int pmf_accessed = min(max_gt, index + seeding_time - 1);
#>   real new_inf = dot_product(infections[inf_start:inf_end], tail(gt_pmf, pmf_accessed));
#>   return(new_inf);
#> }
#> 
#> vector generate_infections(vector R, int seeding_time, real[] gt_mean, 
#>                            real[] gt_sd, int max_gt, vector shifted_cases,
#>                            vector initial_infections) {
#>   int rt = num_elements(R);
#>   int t = rt + seeding_time;
#>   vector[t] infections = rep_vector(1e-5, t);
#>   vector[max_gt] gt_pmf;              // reversed generation time pdf
#>   vector[rt] infectiousness = rep_vector(1e-5, rt);  // infections over time
#>   
#>   for (j in 1:(max_gt)) {
#>     gt_pmf[j] = discretised_gamma_pmf(max_gt - j + 1, gt_mean[1], gt_sd[1], max_gt);
#>   }
#>   // estimate initial infections not using Rt
#>   infections[1:seeding_time] = infections[1:seeding_time] + shifted_cases[1:seeding_time] .* initial_infections;
#>   
#>   for (s in 1:rt) {
#>     infectiousness[s] += update_infectiousness(infections, gt_pmf, seeding_time, max_gt, s);
#>     infections[s + seeding_time] += R[s] * infectiousness[s];
#>   }
#>   return(infections);
#> }
#> 
#> 
#> vector deconvolve_infections(vector shifted_cases, vector noise, int fixed) {
#>   int t = num_elements(shifted_cases);
#>   vector[t] infections = rep_vector(1e-5, t);
#>   if(!fixed) {
#>     infections = infections + shifted_cases .* exp(noise);
#>   }else{
#>     infections = infections + shifted_cases;
#>   }
#>   return(infections);
#> }
#> 
#> void generation_time_lp(real[] gt_mean, real gt_mean_mean, real gt_mean_sd, 
#>                         real[] gt_sd, real gt_sd_mean, real gt_sd_sd, int weight) {
#>     target += normal_lpdf(gt_mean[1] | gt_mean_mean, gt_mean_sd) * weight;
#>     target += normal_lpdf(gt_sd[1] | gt_sd_mean, gt_sd_sd) * weight;
#> }
#> 
#> vector calculate_Rt(vector infections, int seeding_time,
#>                     real gt_mean, real gt_sd, int max_gt) {
#>   vector[max_gt] gt_pmf;  
#>   int t = num_elements(infections);
#>   int ot = t - seeding_time;
#>   vector[ot] R;
#>   vector[ot] infectiousness = rep_vector(1e-5, ot); 
#>   // calculate PMF of the generation time
#>   for (j in 1:(max_gt)) {
#>     gt_pmf[j] =  discretised_gamma_pmf(max_gt - j + 1, gt_mean, gt_sd, max_gt);
#>   }
#>   // calculate Rt using Cori et al. approach
#>   for (s in 1:ot) {
#>     infectiousness[s] += update_infectiousness(infections, gt_pmf, seeding_time, max_gt, s);
#>     R[s] = infections[s + seeding_time] / infectiousness[s];
#>   }
#>   return(R);
#> }
#> 
#> 
#> vector day_of_week_effect(vector reports, int[] day_of_week, vector effect) {
#>   int t = num_elements(reports);
#>   // scale day of week effect
#>   vector[7] scaled_effect = 7 * effect;
#>   vector[t] scaled_reports;
#>   for (s in 1:t) {
#>     // add reporting effects (adjust for simplex scale)
#>     scaled_reports[s] = reports[s] * scaled_effect[day_of_week[s]];
#>    }
#>   return(scaled_reports);
#> }
#> 
#>    
#> void report_lp(int[] cases, vector reports, 
#>                real[] rep_phi, int phi_prior,
#>                int model_type, int horizon,
#>                int weight) {
#>   int t = num_elements(reports) - horizon;
#>   if (model_type) {
#>     //overdispersion
#>     rep_phi[model_type] ~ exponential(phi_prior);
#>     target += neg_binomial_2_lpmf(cases | reports[1:t], rep_phi[model_type]) * weight;
#>   }else{
#>     target += poisson_lpmf(cases | reports[1:t]) * weight;
#>   }
#> }
#> 
#> 
#> real[] R_to_growth(vector R, real gt_mean, real gt_sd) {
#>   real k = pow(gt_sd / gt_mean, 2);
#>   int t = num_elements(R);
#>   real r[t];
#>   for (s in 1:t) {
#>     r[s] = (pow(R[s], k) - 1) / (k * gt_mean);
#>   } 
#>   return(r);
#> }
#> 
#> int[] report_rng(vector reports, real[] rep_phi, int model_type) {
#>   int t = num_elements(reports);
#>   int sampled_reports[t];
#>   if (model_type) {
#>     for (s in 1:t) {
#>       sampled_reports[s] = neg_binomial_2_rng(reports[s] > 1e8 ? 1e8 : reports[s], rep_phi[model_type]);
#>     }
#>   }else{
#>     for (s in 1:t) {
#>       sampled_reports[s] = poisson_rng(reports[s] > 1e8 ? 1e8 : reports[s]);
#>     }
#>   }
#>   return(sampled_reports);
#> }
#> 
#> 
#> }
#> 
#> 
#> data {
#> 
#>   int t;                                            // unobserved time
#>   int seeding_time;                                 // time period used for seeding and not observed
#>   int horizon;                                      // forecast horizon
#>   int<lower = 0> cases[t - horizon - seeding_time]; // observed cases
#>   vector<lower = 0>[t] shifted_cases;               // median shifted smoothed cases
#> 
#>   int delays;                  // no. of delay distributions
#>   real delay_mean_sd[delays];  // prior sd of mean incubation period
#>   real delay_mean_mean[delays];// prior mean of mean incubation period
#>   real delay_sd_mean[delays];  // prior sd of sd of incubation period
#>   real delay_sd_sd[delays];    // prior sd of sd of incubation period
#>   int max_delay[delays];       // maximum incubation period
#> 
#>   real L;				                     // boundary value for infections gp
#>   int<lower=1> M;			               // basis functions for infections gp
#>   real lengthscale_alpha;            // alpha for gp lengthscale prior
#>   real lengthscale_beta;             // beta for gp lengthscale prior
#>   real alpha_sd;                     // standard deviation of the alpha gp kernal parameter
#>   int stationary;                    // is underlying gaussian process first or second order
#>   int fixed;                         //  should a gaussian process be used
#> 
#>   real gt_mean_sd;                   // prior sd of mean generation time
#>   real gt_mean_mean;                 // prior mean of mean generation time
#>   real gt_sd_mean;                   // prior sd of sd of generation time
#>   real gt_sd_sd;                     // prior sd of sd of generation time
#>   int max_gt;                        // maximum generation time
#> 
#>   int estimate_r;                    // should the reproduction no be estimated (1 = yes)
#>   real <lower = 0> r_mean;           // prior mean of reproduction number
#>   real <lower = 0> r_sd;             // prior standard deviation of reproduction number
#>   int bp_n;                          // no of breakpoints (0 = no breakpoints)
#>   int breakpoints[t - seeding_time]; // when do breakpoints occur 
#>   int future_fixed;                  // is underlying future Rt assumed to be fixed
#>   int fixed_from;                    // Reference date for when Rt estimation should be fixed
#> 
#>   int day_of_week[t - seeding_time]; // day of the week indicator (1 - 7)
#>   int model_type;                    // type of model: 0 = poisson otherwise negative binomial
#>   int week_effect;                   // should a day of the week effect be estimated
#> }
#> 
#> transformed data{
#>   // observations
#>   int ot = t - seeding_time - horizon;  // observed time
#>   int ot_h = ot + horizon;  // observed time + forecast horizon
#>   // gaussian process
#>   int noise_terms = setup_noise(ot_h, t, horizon, estimate_r, stationary, future_fixed, fixed_from);
#>   matrix[noise_terms, M] PHI = setup_gp(M, L, noise_terms);  // basis function 
#>   // Rt
#>   real r_logmean = log(r_mean^2 / sqrt(r_sd^2 + r_mean^2)); 
#>   real r_logsd = sqrt(log(1 + (r_sd^2 / r_mean^2))); 
#> }
#> 
#> parameters{
#>   // gaussian process
#>   real<lower = 0> rho[fixed ? 0 : 1];    // length scale of noise GP
#>   real<lower = 0> alpha[fixed ? 0 : 1];  // scale of of noise GP
#>   vector[fixed ? 0 : M] eta;             // unconstrained noise
#>   // Rt
#>   vector[estimate_r] log_R;             // baseline reproduction number estimate (log)
#>   vector[estimate_r > 0 ? seeding_time : 0] initial_infections;// seed infections 
#>   real<lower = 0> gt_mean[estimate_r];  // mean of generation time
#>   real <lower = 0> gt_sd[estimate_r];   // sd of generation time
#>   real bp_effects[bp_n];                // Rt breakpoint effects
#>   // observation model
#>   real<lower = 0> delay_mean[delays];   // mean of delays
#>   real<lower = 0> delay_sd[delays];     // sd of delays
#>   simplex[week_effect ? 7 : 1] day_of_week_simplex;   // day of week reporting effect 
#>   real<lower = 0> rep_phi[model_type];  // overdispersion of the reporting process
#> }
#> 
#> transformed parameters {
#>   vector[fixed ? 0 : noise_terms] noise;                  // noise  generated by the gaussian process
#>   vector[estimate_r > 0 ? ot_h : 0] R;                    // reproduction number
#>   vector[t] infections;                                   // latent infections
#>   vector[ot_h] reports;                                   // observed cases
#>   // GP in noise - spectral densities
#>   if (!fixed) {
#>     noise = update_gp(PHI, M, L, alpha[1], rho[1], eta);
#>   }
#>   // Estimate latent infections
#>   if (estimate_r) {
#>     // via Rt
#>     R = update_Rt(R, log_R[estimate_r], noise, breakpoints, bp_effects, stationary);
#>     infections = generate_infections(R, seeding_time, gt_mean, gt_sd, max_gt, shifted_cases, initial_infections);
#>   }else{
#>     // via deconvolution
#>     infections = deconvolve_infections(shifted_cases, noise, fixed);
#>   }
#>   // convolve from latent infections to mean of observations
#>   reports = convolve_to_report(infections, delay_mean, delay_sd, max_delay, seeding_time);
#>  // weekly reporting effect
#>  if (week_effect) {
#>    reports = day_of_week_effect(reports, day_of_week, day_of_week_simplex);
#>   }
#> }
#> 
#> model {
#>   // priors for noise GP
#>   if (!fixed) {
#>     gaussian_process_lp(rho, alpha, eta, lengthscale_alpha, lengthscale_beta, alpha_sd);
#>   }
#>   // penalised priors for delay distributions
#>   delays_lp(delay_mean, delay_mean_mean, delay_mean_sd, delay_sd, delay_sd_mean, delay_sd_sd, t);
#>   // Rt priors
#>   if (estimate_r) {
#>     // prior on R
#>     log_R ~ normal(r_logmean, r_logsd);
#>     //breakpoint effects on Rt
#>     if (bp_n > 0) {
#>       bp_effects ~ normal(0, 0.1);
#>     }
#>     // initial infections
#>     initial_infections ~ lognormal(0, 0.1);
#>     // penalised_prior on generation interval
#>     generation_time_lp(gt_mean, gt_mean_mean, gt_mean_sd, gt_sd, gt_sd_mean, gt_sd_sd, ot);
#>   }
#>   // observed reports from mean of reports
#>   report_lp(cases, reports, rep_phi, 1, model_type, horizon, 1);
#> }
#>   
#> generated quantities {
#>   int imputed_reports[ot_h]; 
#>   vector[estimate_r > 0 ? 0: ot_h] gen_R;
#>   real r[ot_h];
#>   if (estimate_r){
#>     // estimate growth from estimated Rt
#>     r = R_to_growth(R, gt_mean[1], gt_sd[1]);
#>   }else{
#>     // sample generation time
#>     real gt_mean_sample = normal_rng(gt_mean_mean, gt_mean_sd);
#>     real gt_sd_sample = normal_rng(gt_sd_mean, gt_sd_sd);
#>     // calculate Rt using infections and generation time
#>     gen_R = calculate_Rt(infections, seeding_time, gt_mean_sample, gt_mean_sample, max_gt);
#>     // estimate growth from calculated Rt
#>     r = R_to_growth(gen_R, gt_mean_sample, gt_sd_sample);
#>   }
#>   // simulate reported cases
#>   imputed_reports = report_rng(reports, rep_phi, model_type);
#> } 
#> 
#> $data
#> NULL
#> 
#> $init
#> [1] "random"
#> 
#> $refresh
#> [1] 0
#> 
#> $trials
#> [1] 10
#> 
#> $iter
#> [1] 10000
#> 
#> $output_samples
#> [1] 1000
#> 

# increasing warmup
create_stan_args(stan_args = list(warmup = 1000))

#> $object
#> S4 class stanmodel 'estimate_infections' coded as follows:
#> functions {
#> 
#>  // discretised truncated gamma pmf
#>   real discretised_gamma_pmf(int y, real mu, real sigma, int max_val) {
#>     // calculate alpha and beta for gamma distribution
#>     real c_sigma = sigma + 1e-5;
#>     real alpha = ((mu)/ c_sigma)^2;
#>     real beta = (mu) / (c_sigma^2);
#>     //account for numerical issues
#>     alpha = alpha <= 0 ? 1e-5 : alpha;
#>     beta = beta <= 0 ? 1e-5 : beta;
#>     alpha = is_inf(alpha) ? 1e8 : alpha;
#>     beta = is_inf(beta) ? 1e8 : beta; 
#>     return((gamma_cdf(y + 1, alpha, beta) - gamma_cdf(y, alpha, beta)) / 
#>     (gamma_cdf(max_val + 1, alpha, beta) - gamma_cdf(1, alpha, beta)));
#>   }
#> 
#> 
#>   // discretised truncated lognormal pmf
#>   real discretised_lognormal_pmf(int y, real mu, real sigma, int max_val) {
#>     real small = 1e-5;
#>     real adj_y = y + small;
#>     return((normal_cdf((log(adj_y + 1) - mu) / sigma, 0.0, 1.0) -
#>             normal_cdf((log(adj_y) - mu) / sigma, 0.0, 1.0)) /
#>            (normal_cdf((log(max_val + small) - mu) / sigma, 0.0, 1.0) -
#>             normal_cdf((log(small) - mu) / sigma, 0.0, 1.0)));
#>   }
#> 
#> // convolve a pdf and case vector 
#> vector convolve(vector cases, vector pdf) {
#>     int t = num_elements(cases);
#>     int max_pdf = num_elements(pdf);
#>     vector[t] convolved_cases = rep_vector(1e-5, t);
#>     for (s in 1:t) {
#>         convolved_cases[s] += dot_product(cases[max(1, (s - max_pdf + 1)):s], tail(pdf, min(max_pdf, s)));
#>     }
#>    return(convolved_cases);
#>   }
#> 
#> // convolve latent infections to reported (but still unobserved) cases
#> vector convolve_to_report(vector infections, 
#>                           real[] delay_mean, 
#>                           real[] delay_sd,
#>                           int[] max_delay,
#>                           int seeding_time) {
#>   int t = num_elements(infections);
#>   vector[t - seeding_time] reports;
#>   vector[t] reports_hold = infections;
#>   int delays = num_elements(delay_mean);
#>   if (delays) {
#>     for (s in 1:delays) {
#>       vector[max_delay[s]] rev_delay = rep_vector(1e-5, max_delay[s]);
#>       for (j in 1:(max_delay[s])) {
#>         rev_delay[j] +=
#>         discretised_lognormal_pmf(max_delay[s] - j, delay_mean[s], delay_sd[s], max_delay[s]);
#>       }
#>       reports_hold = convolve(reports_hold, rev_delay);
#>     }
#>     reports = reports_hold[(seeding_time + 1):t];
#>   }else{
#>     reports = infections[(seeding_time + 1):t];
#>   }
#>   return(reports);
#> }
#> 
#> void delays_lp(real[] delay_mean, real[] delay_mean_mean, real[] delay_mean_sd, 
#>                real[] delay_sd, real[] delay_sd_mean, real[] delay_sd_sd, int weight){
#>     int delays = num_elements(delay_mean);
#>     if (delays) {
#>     for (s in 1:delays) {
#>       target += normal_lpdf(delay_mean[s] | delay_mean_mean[s], delay_mean_sd[s]) * weight;
#>       target += normal_lpdf(delay_sd[s] | delay_sd_mean[s], delay_sd_sd[s]) * weight;
#>     }
#>   }
#> }
#> 
#>   // exponential quadratic kernal
#> 	real spd_SE(real alpha, real rho, real w) {
#> 		real S;
#> 		S = (alpha^2) * sqrt(2*pi()) * rho * exp(-0.5*(rho^2)*(w^2));
#> 		return S;
#> 	}
#> 	
#> 	// basis function for approximate hilbert space gp
#> 	// see here for details: https://arxiv.org/pdf/2004.11408.pdf
#> 	vector phi_SE(real L, int m, vector x) {
#> 		vector[rows(x)] fi;
#> 		fi = 1/sqrt(L) * sin(m*pi()/(2*L) * (x+L));
#> 		return fi;
#> 	}
#> 	
#> 	// eigenvalues for approximate hilbert space gp
#> 	// see here for details: https://arxiv.org/pdf/2004.11408.pdf
#> 	real lambda(real L, int m) {
#> 		real lam;
#> 		lam = ((m*pi())/(2*L))^2;
#> 		return lam;
#> 	}
#> 
#> 
#> int setup_noise(int ot_h, int t, int horizon, int estimate_r, 
#>                 int stationary, int future_fixed, int fixed_from) {
#>   int noise_time = estimate_r > 0 ? (stationary > 0 ? ot_h : ot_h - 1) : t;
#>   int noise_terms =  future_fixed > 0 ? (noise_time - horizon + fixed_from) : noise_time; 
#>   return(noise_terms);
#> }
#> 
#> 
#> matrix setup_gp(int M, real L, int dimension) {
#>   vector[dimension] time;
#>   matrix[dimension, M] PHI;
#>   for (s in 1:dimension) {
#>     time[s] = s;
#>   }
#>   for (m in 1:M){ 
#>     PHI[,m] = phi_SE(L, m, time); 
#>   }
#>   return(PHI);
#> }
#> 
#> 
#> vector update_gp(matrix PHI, int M, real L, real alpha, real rho, vector eta) {
#>   vector[M] diagSPD;    // spectral density
#>   vector[M] SPD_eta;    // spectral density * noise
#>   int noise_terms = rows(PHI);
#>   vector[noise_terms] noise = rep_vector(1e-6, noise_terms);
#>   // GP in noise - spectral densities
#>   for(m in 1:M){ 
#>     diagSPD[m] =  sqrt(spd_SE(alpha, rho, sqrt(lambda(L, m)))); 
#>   }
#>   SPD_eta = diagSPD .* eta;
#>   noise = noise + PHI[,] * SPD_eta;
#>   return(noise);
#> }
#> 
#> void gaussian_process_lp(real[] rho, real[] alpha, vector eta,
#>                          real ls_alpha, real ls_beta, real alpha_sd) {
#>   rho ~ inv_gamma(ls_alpha, ls_beta);
#>   alpha ~ normal(0, alpha_sd);
#>   eta ~ std_normal();
#> }
#> 
#> 
#> 
#> real update_breakpoints(real input_R, real[] bp_effects,
#>                         int bp_index, int at_bp,
#>                         int stationary) {
#>   real R = input_R;
#>   if (stationary) {
#>     if (bp_index > 0) {
#>       R += sum(bp_effects[1:bp_index]);
#>       }
#>   }else{
#>     if (at_bp) {
#>       R += bp_effects[bp_index];
#>     }
#>   }               
#>   return(R);
#> }
#> 
#> real update_R(vector R, vector noise, int noise_terms,
#>               int index, int stationary) {
#>   real cR = R[index];
#>   if (noise_terms > 0) {
#>     if (stationary){
#>       if (index <= noise_terms) {
#>         cR += noise[index];
#>       }else{
#>         if (index > 1) {
#>            cR = R[index - 1];
#>         }
#>       }
#>     }else{
#>       if (index <= (noise_terms + 1)) {
#>         cR = R[index - 1] + noise[index - 1];
#>       }else{
#>         cR = R[index - 1];
#>       }
#>     }
#>   }
#>   return(cR);
#> }
#> 
#> vector update_Rt(vector input_R, real log_R, vector noise, int[] bps,
#>                  real[] bp_effects, int stationary) {
#>   // define control parameters
#>   int noise_terms = num_elements(noise);
#>   int i_stationary = noise_terms > 0 ? stationary : 1;
#>   int t = num_elements(input_R);
#>   int bp_n = num_elements(bp_effects);
#>   int bp_in = 0;
#>   int at_bp = 0;
#>   int index;
#>   vector[t] R;
#>   // initialise Rt
#>   if (i_stationary) {
#>     R = rep_vector(log_R, t);
#>     index = 1;
#>   }else{
#>     R[1] = log_R;
#>     index = 2;
#>   }
#>   // iteratively update Rt
#>   for (s in index:t) {
#>     R[s] = update_R(R, noise, noise_terms, s, i_stationary);
#>     if (bp_n > 0) {
#>       at_bp = bps[s];
#>       bp_in += at_bp;
#>       R[s] = update_breakpoints(R[s], bp_effects, bp_in, at_bp, i_stationary);
#>     }
#>   }
#>   // convert to correct scale
#>   R = exp(R);
#>   return(R);
#> }
#> 
#> 
#> real update_infectiousness(vector infections, vector gt_pmf,
#>                            int seeding_time, int max_gt, int index){
#>   int inf_start = max(1, (index + seeding_time - max_gt));
#>   int inf_end = (index + seeding_time - 1);
#>   int pmf_accessed = min(max_gt, index + seeding_time - 1);
#>   real new_inf = dot_product(infections[inf_start:inf_end], tail(gt_pmf, pmf_accessed));
#>   return(new_inf);
#> }
#> 
#> vector generate_infections(vector R, int seeding_time, real[] gt_mean, 
#>                            real[] gt_sd, int max_gt, vector shifted_cases,
#>                            vector initial_infections) {
#>   int rt = num_elements(R);
#>   int t = rt + seeding_time;
#>   vector[t] infections = rep_vector(1e-5, t);
#>   vector[max_gt] gt_pmf;              // reversed generation time pdf
#>   vector[rt] infectiousness = rep_vector(1e-5, rt);  // infections over time
#>   
#>   for (j in 1:(max_gt)) {
#>     gt_pmf[j] = discretised_gamma_pmf(max_gt - j + 1, gt_mean[1], gt_sd[1], max_gt);
#>   }
#>   // estimate initial infections not using Rt
#>   infections[1:seeding_time] = infections[1:seeding_time] + shifted_cases[1:seeding_time] .* initial_infections;
#>   
#>   for (s in 1:rt) {
#>     infectiousness[s] += update_infectiousness(infections, gt_pmf, seeding_time, max_gt, s);
#>     infections[s + seeding_time] += R[s] * infectiousness[s];
#>   }
#>   return(infections);
#> }
#> 
#> 
#> vector deconvolve_infections(vector shifted_cases, vector noise, int fixed) {
#>   int t = num_elements(shifted_cases);
#>   vector[t] infections = rep_vector(1e-5, t);
#>   if(!fixed) {
#>     infections = infections + shifted_cases .* exp(noise);
#>   }else{
#>     infections = infections + shifted_cases;
#>   }
#>   return(infections);
#> }
#> 
#> void generation_time_lp(real[] gt_mean, real gt_mean_mean, real gt_mean_sd, 
#>                         real[] gt_sd, real gt_sd_mean, real gt_sd_sd, int weight) {
#>     target += normal_lpdf(gt_mean[1] | gt_mean_mean, gt_mean_sd) * weight;
#>     target += normal_lpdf(gt_sd[1] | gt_sd_mean, gt_sd_sd) * weight;
#> }
#> 
#> vector calculate_Rt(vector infections, int seeding_time,
#>                     real gt_mean, real gt_sd, int max_gt) {
#>   vector[max_gt] gt_pmf;  
#>   int t = num_elements(infections);
#>   int ot = t - seeding_time;
#>   vector[ot] R;
#>   vector[ot] infectiousness = rep_vector(1e-5, ot); 
#>   // calculate PMF of the generation time
#>   for (j in 1:(max_gt)) {
#>     gt_pmf[j] =  discretised_gamma_pmf(max_gt - j + 1, gt_mean, gt_sd, max_gt);
#>   }
#>   // calculate Rt using Cori et al. approach
#>   for (s in 1:ot) {
#>     infectiousness[s] += update_infectiousness(infections, gt_pmf, seeding_time, max_gt, s);
#>     R[s] = infections[s + seeding_time] / infectiousness[s];
#>   }
#>   return(R);
#> }
#> 
#> 
#> vector day_of_week_effect(vector reports, int[] day_of_week, vector effect) {
#>   int t = num_elements(reports);
#>   // scale day of week effect
#>   vector[7] scaled_effect = 7 * effect;
#>   vector[t] scaled_reports;
#>   for (s in 1:t) {
#>     // add reporting effects (adjust for simplex scale)
#>     scaled_reports[s] = reports[s] * scaled_effect[day_of_week[s]];
#>    }
#>   return(scaled_reports);
#> }
#> 
#>    
#> void report_lp(int[] cases, vector reports, 
#>                real[] rep_phi, int phi_prior,
#>                int model_type, int horizon,
#>                int weight) {
#>   int t = num_elements(reports) - horizon;
#>   if (model_type) {
#>     //overdispersion
#>     rep_phi[model_type] ~ exponential(phi_prior);
#>     target += neg_binomial_2_lpmf(cases | reports[1:t], rep_phi[model_type]) * weight;
#>   }else{
#>     target += poisson_lpmf(cases | reports[1:t]) * weight;
#>   }
#> }
#> 
#> 
#> real[] R_to_growth(vector R, real gt_mean, real gt_sd) {
#>   real k = pow(gt_sd / gt_mean, 2);
#>   int t = num_elements(R);
#>   real r[t];
#>   for (s in 1:t) {
#>     r[s] = (pow(R[s], k) - 1) / (k * gt_mean);
#>   } 
#>   return(r);
#> }
#> 
#> int[] report_rng(vector reports, real[] rep_phi, int model_type) {
#>   int t = num_elements(reports);
#>   int sampled_reports[t];
#>   if (model_type) {
#>     for (s in 1:t) {
#>       sampled_reports[s] = neg_binomial_2_rng(reports[s] > 1e8 ? 1e8 : reports[s], rep_phi[model_type]);
#>     }
#>   }else{
#>     for (s in 1:t) {
#>       sampled_reports[s] = poisson_rng(reports[s] > 1e8 ? 1e8 : reports[s]);
#>     }
#>   }
#>   return(sampled_reports);
#> }
#> 
#> 
#> }
#> 
#> 
#> data {
#> 
#>   int t;                                            // unobserved time
#>   int seeding_time;                                 // time period used for seeding and not observed
#>   int horizon;                                      // forecast horizon
#>   int<lower = 0> cases[t - horizon - seeding_time]; // observed cases
#>   vector<lower = 0>[t] shifted_cases;               // median shifted smoothed cases
#> 
#>   int delays;                  // no. of delay distributions
#>   real delay_mean_sd[delays];  // prior sd of mean incubation period
#>   real delay_mean_mean[delays];// prior mean of mean incubation period
#>   real delay_sd_mean[delays];  // prior sd of sd of incubation period
#>   real delay_sd_sd[delays];    // prior sd of sd of incubation period
#>   int max_delay[delays];       // maximum incubation period
#> 
#>   real L;				                     // boundary value for infections gp
#>   int<lower=1> M;			               // basis functions for infections gp
#>   real lengthscale_alpha;            // alpha for gp lengthscale prior
#>   real lengthscale_beta;             // beta for gp lengthscale prior
#>   real alpha_sd;                     // standard deviation of the alpha gp kernal parameter
#>   int stationary;                    // is underlying gaussian process first or second order
#>   int fixed;                         //  should a gaussian process be used
#> 
#>   real gt_mean_sd;                   // prior sd of mean generation time
#>   real gt_mean_mean;                 // prior mean of mean generation time
#>   real gt_sd_mean;                   // prior sd of sd of generation time
#>   real gt_sd_sd;                     // prior sd of sd of generation time
#>   int max_gt;                        // maximum generation time
#> 
#>   int estimate_r;                    // should the reproduction no be estimated (1 = yes)
#>   real <lower = 0> r_mean;           // prior mean of reproduction number
#>   real <lower = 0> r_sd;             // prior standard deviation of reproduction number
#>   int bp_n;                          // no of breakpoints (0 = no breakpoints)
#>   int breakpoints[t - seeding_time]; // when do breakpoints occur 
#>   int future_fixed;                  // is underlying future Rt assumed to be fixed
#>   int fixed_from;                    // Reference date for when Rt estimation should be fixed
#> 
#>   int day_of_week[t - seeding_time]; // day of the week indicator (1 - 7)
#>   int model_type;                    // type of model: 0 = poisson otherwise negative binomial
#>   int week_effect;                   // should a day of the week effect be estimated
#> }
#> 
#> transformed data{
#>   // observations
#>   int ot = t - seeding_time - horizon;  // observed time
#>   int ot_h = ot + horizon;  // observed time + forecast horizon
#>   // gaussian process
#>   int noise_terms = setup_noise(ot_h, t, horizon, estimate_r, stationary, future_fixed, fixed_from);
#>   matrix[noise_terms, M] PHI = setup_gp(M, L, noise_terms);  // basis function 
#>   // Rt
#>   real r_logmean = log(r_mean^2 / sqrt(r_sd^2 + r_mean^2)); 
#>   real r_logsd = sqrt(log(1 + (r_sd^2 / r_mean^2))); 
#> }
#> 
#> parameters{
#>   // gaussian process
#>   real<lower = 0> rho[fixed ? 0 : 1];    // length scale of noise GP
#>   real<lower = 0> alpha[fixed ? 0 : 1];  // scale of of noise GP
#>   vector[fixed ? 0 : M] eta;             // unconstrained noise
#>   // Rt
#>   vector[estimate_r] log_R;             // baseline reproduction number estimate (log)
#>   vector[estimate_r > 0 ? seeding_time : 0] initial_infections;// seed infections 
#>   real<lower = 0> gt_mean[estimate_r];  // mean of generation time
#>   real <lower = 0> gt_sd[estimate_r];   // sd of generation time
#>   real bp_effects[bp_n];                // Rt breakpoint effects
#>   // observation model
#>   real<lower = 0> delay_mean[delays];   // mean of delays
#>   real<lower = 0> delay_sd[delays];     // sd of delays
#>   simplex[week_effect ? 7 : 1] day_of_week_simplex;   // day of week reporting effect 
#>   real<lower = 0> rep_phi[model_type];  // overdispersion of the reporting process
#> }
#> 
#> transformed parameters {
#>   vector[fixed ? 0 : noise_terms] noise;                  // noise  generated by the gaussian process
#>   vector[estimate_r > 0 ? ot_h : 0] R;                    // reproduction number
#>   vector[t] infections;                                   // latent infections
#>   vector[ot_h] reports;                                   // observed cases
#>   // GP in noise - spectral densities
#>   if (!fixed) {
#>     noise = update_gp(PHI, M, L, alpha[1], rho[1], eta);
#>   }
#>   // Estimate latent infections
#>   if (estimate_r) {
#>     // via Rt
#>     R = update_Rt(R, log_R[estimate_r], noise, breakpoints, bp_effects, stationary);
#>     infections = generate_infections(R, seeding_time, gt_mean, gt_sd, max_gt, shifted_cases, initial_infections);
#>   }else{
#>     // via deconvolution
#>     infections = deconvolve_infections(shifted_cases, noise, fixed);
#>   }
#>   // convolve from latent infections to mean of observations
#>   reports = convolve_to_report(infections, delay_mean, delay_sd, max_delay, seeding_time);
#>  // weekly reporting effect
#>  if (week_effect) {
#>    reports = day_of_week_effect(reports, day_of_week, day_of_week_simplex);
#>   }
#> }
#> 
#> model {
#>   // priors for noise GP
#>   if (!fixed) {
#>     gaussian_process_lp(rho, alpha, eta, lengthscale_alpha, lengthscale_beta, alpha_sd);
#>   }
#>   // penalised priors for delay distributions
#>   delays_lp(delay_mean, delay_mean_mean, delay_mean_sd, delay_sd, delay_sd_mean, delay_sd_sd, t);
#>   // Rt priors
#>   if (estimate_r) {
#>     // prior on R
#>     log_R ~ normal(r_logmean, r_logsd);
#>     //breakpoint effects on Rt
#>     if (bp_n > 0) {
#>       bp_effects ~ normal(0, 0.1);
#>     }
#>     // initial infections
#>     initial_infections ~ lognormal(0, 0.1);
#>     // penalised_prior on generation interval
#>     generation_time_lp(gt_mean, gt_mean_mean, gt_mean_sd, gt_sd, gt_sd_mean, gt_sd_sd, ot);
#>   }
#>   // observed reports from mean of reports
#>   report_lp(cases, reports, rep_phi, 1, model_type, horizon, 1);
#> }
#>   
#> generated quantities {
#>   int imputed_reports[ot_h]; 
#>   vector[estimate_r > 0 ? 0: ot_h] gen_R;
#>   real r[ot_h];
#>   if (estimate_r){
#>     // estimate growth from estimated Rt
#>     r = R_to_growth(R, gt_mean[1], gt_sd[1]);
#>   }else{
#>     // sample generation time
#>     real gt_mean_sample = normal_rng(gt_mean_mean, gt_mean_sd);
#>     real gt_sd_sample = normal_rng(gt_sd_mean, gt_sd_sd);
#>     // calculate Rt using infections and generation time
#>     gen_R = calculate_Rt(infections, seeding_time, gt_mean_sample, gt_mean_sample, max_gt);
#>     // estimate growth from calculated Rt
#>     r = R_to_growth(gen_R, gt_mean_sample, gt_sd_sample);
#>   }
#>   // simulate reported cases
#>   imputed_reports = report_rng(reports, rep_phi, model_type);
#> } 
#> 
#> $data
#> NULL
#> 
#> $init
#> [1] "random"
#> 
#> $refresh
#> [1] 0
#> 
#> $cores
#> [1] 4
#> 
#> $chains
#> [1] 4
#> 
#> $control
#> $control$adapt_delta
#> [1] 0.99
#> 
#> $control$max_treedepth
#> [1] 15
#> 
#> 
#> $save_warmup
#> [1] FALSE
#> 
#> $seed
#> [1] 15814683
#> 
#> $warmup
#> [1] 1000
#> 
#> $iter
#> [1] 1250
#>