Delay Distribution Options

Returns delay distributions formatted for usage by downstream functions.

Usage

delay_opts(..., fixed = FALSE)

Arguments

...

A list of lists specifying distributions of reporting delays. Each list is passed to dist_spec() and so should contain parameters linked to the functions arguments. If a list is not given then it is assumed that no delay is present. If multiple lists are given then they are assumed to be independent.

fixed

Logical, defaults to FALSE. Should all reporting delays be treated as coming from fixed (vs uncertain) distributions. Making this simplification drastically reduces compute requirements. Setting this here overrides any of the constituent delay distributions being set to be fixed or not.

Value

A list summarising the input delay distributions.

See also

Author

Sam Abbott

Sebastian Funk

Examples

# no delays
delay_opts()
#> $delay_mean_mean
#> numeric(0)
#> 
#> $delay_mean_sd
#> numeric(0)
#> 
#> $delay_sd_mean
#> numeric(0)
#> 
#> $delay_sd_sd
#> numeric(0)
#> 
#> $delay_fixed
#> integer(0)
#> 
#> $delay_dist
#> integer(0)
#> 
#> $delay_max
#> integer(0)
#> 
#> $seeding_time
#> [1] 1
#> 
#> $delays
#> [1] 0
#> 
#> $uncertain_mean_delays
#> integer(0)
#> 
#> $uncertain_sd_delays
#> integer(0)
#> 
#> $fixed_delays
#> integer(0)
#> 
#> $n_uncertain_mean_delays
#> [1] 0
#> 
#> $n_uncertain_sd_delays
#> [1] 0
#> 
#> $n_fixed_delays
#> [1] 0
#> 

# A single delay that has uncertainty
delay <- list(mean = 1, mean_sd = 0.2, sd = 0.5, sd_sd = 0.1, max = 15)
delay_opts(delay)
#> $delay_mean_mean
#> [1] 1
#> 
#> $delay_mean_sd
#> [1] 0.2
#> 
#> $delay_sd_mean
#> [1] 0.5
#> 
#> $delay_sd_sd
#> [1] 0.1
#> 
#> $delay_fixed
#> [1] 0
#> 
#> $delay_max
#> [1] 15
#> 
#> $delay_dist
#> [1] 0
#> 
#> $seeding_time
#> [1] 3
#> 
#> $delays
#> [1] 1
#> 
#> $uncertain_mean_delays
#> [1] 1
#> 
#> $uncertain_sd_delays
#> [1] 1
#> 
#> $fixed_delays
#> integer(0)
#> 
#> $n_uncertain_mean_delays
#> [1] 1
#> 
#> $n_uncertain_sd_delays
#> [1] 1
#> 
#> $n_fixed_delays
#> [1] 0
#> 

# A single delay where we override the uncertainty assumption
delay_opts(delay, fixed = TRUE)
#> $delay_mean_mean
#> [1] 1
#> 
#> $delay_mean_sd
#> [1] 0
#> 
#> $delay_sd_mean
#> [1] 0.5
#> 
#> $delay_sd_sd
#> [1] 0
#> 
#> $delay_fixed
#> [1] 1
#> 
#> $delay_max
#> [1] 15
#> 
#> $delay_dist
#> [1] 0
#> 
#> $seeding_time
#> [1] 3
#> 
#> $delays
#> [1] 1
#> 
#> $uncertain_mean_delays
#> integer(0)
#> 
#> $uncertain_sd_delays
#> integer(0)
#> 
#> $fixed_delays
#> [1] 1
#> 
#> $n_uncertain_mean_delays
#> [1] 0
#> 
#> $n_uncertain_sd_delays
#> [1] 0
#> 
#> $n_fixed_delays
#> [1] 1
#> 

# A delay where uncertainty is implict
delay_opts(list(mean = 1, mean_sd = 0, sd = 0.5, sd_sd = 0, max = 15))
#> $delay_mean_mean
#> [1] 1
#> 
#> $delay_mean_sd
#> [1] 0
#> 
#> $delay_sd_mean
#> [1] 0.5
#> 
#> $delay_sd_sd
#> [1] 0
#> 
#> $delay_fixed
#> [1] 1
#> 
#> $delay_max
#> [1] 15
#> 
#> $delay_dist
#> [1] 0
#> 
#> $seeding_time
#> [1] 3
#> 
#> $delays
#> [1] 1
#> 
#> $uncertain_mean_delays
#> integer(0)
#> 
#> $uncertain_sd_delays
#> integer(0)
#> 
#> $fixed_delays
#> [1] 1
#> 
#> $n_uncertain_mean_delays
#> [1] 0
#> 
#> $n_uncertain_sd_delays
#> [1] 0
#> 
#> $n_fixed_delays
#> [1] 1
#> 

# Multiple delays
delay_opts(delay, delay)
#> $delay_mean_mean
#> [1] 1 1
#> 
#> $delay_mean_sd
#> [1] 0.2 0.2
#> 
#> $delay_sd_mean
#> [1] 0.5 0.5
#> 
#> $delay_sd_sd
#> [1] 0.1 0.1
#> 
#> $delay_fixed
#> [1] 0 0
#> 
#> $delay_max
#> [1] 15 15
#> 
#> $delay_dist
#> [1] 0 0
#> 
#> $seeding_time
#> [1] 6
#> 
#> $delays
#> [1] 2
#> 
#> $uncertain_mean_delays
#> [1] 1 2
#> 
#> $uncertain_sd_delays
#> [1] 1 2
#> 
#> $fixed_delays
#> integer(0)
#> 
#> $n_uncertain_mean_delays
#> [1] 2
#> 
#> $n_uncertain_sd_delays
#> [1] 2
#> 
#> $n_fixed_delays
#> [1] 0
#>