Forecast Infections and the Time-Varying Reproduction Number

This function provides optional tools for forecasting cases and Rt estimates using the timeseries methods (via the EpiSoon package). It requires the Episoon package. Installation instructions for the EpiSoon package are available here.

forecast_infections(
  infections,
  rts,
  gt_mean,
  gt_sd,
  gt_max = 30,
  ensemble_type = "mean",
  forecast_model,
  horizon = 14,
  samples = 1000
)

Arguments

infections
A data frame of cases by date of infection containing the following variables: date, mean, sd

rts
A data frame of Rt estimates by date of infection containing the following variables: date, mean, sd

gt_mean
Numeric, the mean of the gamma distributed generation time.

gt_sd
Numeric, the standard deviation of the gamma distributed generation time.

gt_max
Numeric, the maximum allowed value of the gamma distributed generation time.

ensemble_type
Character string indicating the type of ensemble to use. By default this is an unweighted ensemble ("mean") with no other types currently supported.

forecast_model
An uninitialised forecast model function to be passed to EpiSoon::forecast_rt. Used for forecasting future Rt and case co An example of the required structure is: function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}.

horizon
Numeric, defaults to 14. The horizon over which to forecast Rts and cases.

samples
Numeric, the number of forecast samples to take.

Value

A list of data.tables. The first entry ("samples") contains raw forecast samples and the second entry ("summarised") contains summarised forecasts.

Examples

# \donttest{

if(requireNamespace("EpiSoon")){
   if(requireNamespace("forecastHybrid")){

## Example case data
reported_cases <- EpiNow2::example_confirmed[1:40]
 
generation_time <- list(mean = EpiNow2::covid_generation_times[1, ]$mean,
                        mean_sd = EpiNow2::covid_generation_times[1, ]$mean_sd,
                        sd = EpiNow2::covid_generation_times[1, ]$sd,
                        sd_sd = EpiNow2::covid_generation_times[1, ]$sd_sd,
                        max = 30)
                          
incubation_period <- list(mean = EpiNow2::covid_incubation_period[1, ]$mean,
                          mean_sd = EpiNow2::covid_incubation_period[1, ]$mean_sd,
                          sd = EpiNow2::covid_incubation_period[1, ]$sd,
                          sd_sd = EpiNow2::covid_incubation_period[1, ]$sd_sd,
                          max = 30)
                   
reporting_delay <- list(mean = log(5),
                        mean_sd = log(2),
                        sd = log(2),
                        sd_sd = log(1.5),
                        max = 30)
                        
## Estimate Rt and infections from data
out <- estimate_infections(reported_cases, family = "negbin",
                           generation_time = generation_time,
                           delays = list(incubation_period, reporting_delay),
                           rt_prior = list(mean = 1, sd = 1),
                           samples = 1000, warmup = 500,
                           cores = ifelse(interactive(), 4, 1), chains = 4,
                           estimate_rt = TRUE,
                           verbose = TRUE, return_fit = TRUE)

## Forecast Rt and infections from estimates
forecast <- forecast_infections(
    infections = out$summarised[variable == "infections"],
    rts = out$summarised[variable == "R"],
     gt_mean = out$summarised[variable == "gt_mean"]$mean,
     gt_sd = out$summarised[variable == "gt_sd"]$mean,
     gt_max = 30,
     forecast_model = function(y, ...){
       EpiSoon::forecastHybrid_model(y = y[max(1, length(y) - 21):length(y)],
       model_params = list(models = "aefz", weights = "equal"),
       forecast_params = list(PI.combination = "mean"), ...)},
     horizon = 14, 
     samples = 1000)
                                
forecast
  }
 }

#> Running for 1000 samples (across 4 chains each with a warm up of 500 iterations each) and 58 time steps of which 7 are a forecast

#> 
#> SAMPLING FOR MODEL 'estimate_infections' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000284 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 2.84 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 750 [  0%]  (Warmup)
#> Chain 1: Iteration:  50 / 750 [  6%]  (Warmup)
#> Chain 1: Iteration: 100 / 750 [ 13%]  (Warmup)
#> Chain 1: Iteration: 150 / 750 [ 20%]  (Warmup)
#> Chain 1: Iteration: 200 / 750 [ 26%]  (Warmup)
#> Chain 1: Iteration: 250 / 750 [ 33%]  (Warmup)
#> Chain 1: Iteration: 300 / 750 [ 40%]  (Warmup)
#> Chain 1: Iteration: 350 / 750 [ 46%]  (Warmup)
#> Chain 1: Iteration: 400 / 750 [ 53%]  (Warmup)
#> Chain 1: Iteration: 450 / 750 [ 60%]  (Warmup)
#> Chain 1: Iteration: 500 / 750 [ 66%]  (Warmup)
#> Chain 1: Iteration: 501 / 750 [ 66%]  (Sampling)
#> Chain 1: Iteration: 550 / 750 [ 73%]  (Sampling)
#> Chain 1: Iteration: 600 / 750 [ 80%]  (Sampling)
#> Chain 1: Iteration: 650 / 750 [ 86%]  (Sampling)
#> Chain 1: Iteration: 700 / 750 [ 93%]  (Sampling)
#> Chain 1: Iteration: 750 / 750 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 110.353 seconds (Warm-up)
#> Chain 1:                35.8133 seconds (Sampling)
#> Chain 1:                146.166 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'estimate_infections' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000297 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 2.97 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:   1 / 750 [  0%]  (Warmup)
#> Chain 2: Iteration:  50 / 750 [  6%]  (Warmup)
#> Chain 2: Iteration: 100 / 750 [ 13%]  (Warmup)
#> Chain 2: Iteration: 150 / 750 [ 20%]  (Warmup)
#> Chain 2: Iteration: 200 / 750 [ 26%]  (Warmup)
#> Chain 2: Iteration: 250 / 750 [ 33%]  (Warmup)
#> Chain 2: Iteration: 300 / 750 [ 40%]  (Warmup)
#> Chain 2: Iteration: 350 / 750 [ 46%]  (Warmup)
#> Chain 2: Iteration: 400 / 750 [ 53%]  (Warmup)
#> Chain 2: Iteration: 450 / 750 [ 60%]  (Warmup)
#> Chain 2: Iteration: 500 / 750 [ 66%]  (Warmup)
#> Chain 2: Iteration: 501 / 750 [ 66%]  (Sampling)
#> Chain 2: Iteration: 550 / 750 [ 73%]  (Sampling)
#> Chain 2: Iteration: 600 / 750 [ 80%]  (Sampling)
#> Chain 2: Iteration: 650 / 750 [ 86%]  (Sampling)
#> Chain 2: Iteration: 700 / 750 [ 93%]  (Sampling)
#> Chain 2: Iteration: 750 / 750 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 108.536 seconds (Warm-up)
#> Chain 2:                35.5715 seconds (Sampling)
#> Chain 2:                144.108 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'estimate_infections' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000329 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 3.29 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:   1 / 750 [  0%]  (Warmup)
#> Chain 3: Iteration:  50 / 750 [  6%]  (Warmup)
#> Chain 3: Iteration: 100 / 750 [ 13%]  (Warmup)
#> Chain 3: Iteration: 150 / 750 [ 20%]  (Warmup)
#> Chain 3: Iteration: 200 / 750 [ 26%]  (Warmup)
#> Chain 3: Iteration: 250 / 750 [ 33%]  (Warmup)
#> Chain 3: Iteration: 300 / 750 [ 40%]  (Warmup)
#> Chain 3: Iteration: 350 / 750 [ 46%]  (Warmup)
#> Chain 3: Iteration: 400 / 750 [ 53%]  (Warmup)
#> Chain 3: Iteration: 450 / 750 [ 60%]  (Warmup)
#> Chain 3: Iteration: 500 / 750 [ 66%]  (Warmup)
#> Chain 3: Iteration: 501 / 750 [ 66%]  (Sampling)
#> Chain 3: Iteration: 550 / 750 [ 73%]  (Sampling)
#> Chain 3: Iteration: 600 / 750 [ 80%]  (Sampling)
#> Chain 3: Iteration: 650 / 750 [ 86%]  (Sampling)
#> Chain 3: Iteration: 700 / 750 [ 93%]  (Sampling)
#> Chain 3: Iteration: 750 / 750 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 109.435 seconds (Warm-up)
#> Chain 3:                35.8114 seconds (Sampling)
#> Chain 3:                145.247 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'estimate_infections' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000453 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 4.53 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:   1 / 750 [  0%]  (Warmup)
#> Chain 4: Iteration:  50 / 750 [  6%]  (Warmup)
#> Chain 4: Iteration: 100 / 750 [ 13%]  (Warmup)
#> Chain 4: Iteration: 150 / 750 [ 20%]  (Warmup)
#> Chain 4: Iteration: 200 / 750 [ 26%]  (Warmup)
#> Chain 4: Iteration: 250 / 750 [ 33%]  (Warmup)
#> Chain 4: Iteration: 300 / 750 [ 40%]  (Warmup)
#> Chain 4: Iteration: 350 / 750 [ 46%]  (Warmup)
#> Chain 4: Iteration: 400 / 750 [ 53%]  (Warmup)
#> Chain 4: Iteration: 450 / 750 [ 60%]  (Warmup)
#> Chain 4: Iteration: 500 / 750 [ 66%]  (Warmup)
#> Chain 4: Iteration: 501 / 750 [ 66%]  (Sampling)
#> Chain 4: Iteration: 550 / 750 [ 73%]  (Sampling)
#> Chain 4: Iteration: 600 / 750 [ 80%]  (Sampling)
#> Chain 4: Iteration: 650 / 750 [ 86%]  (Sampling)
#> Chain 4: Iteration: 700 / 750 [ 93%]  (Sampling)
#> Chain 4: Iteration: 750 / 750 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 110.669 seconds (Warm-up)
#> Chain 4:                35.2232 seconds (Sampling)
#> Chain 4:                145.892 seconds (Total)
#> Chain 4: 

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> $samples
#>        sample       date horizon        value type forecast_type
#>     1:      1 2020-04-09       1 9.519944e-01   rt          <NA>
#>     2:      2 2020-04-09       1 1.459277e+00   rt          <NA>
#>     3:      3 2020-04-09       1 1.273975e+00   rt          <NA>
#>     4:      4 2020-04-09       1 1.248591e+00   rt          <NA>
#>     5:      5 2020-04-09       1 1.604573e+00   rt          <NA>
#>    ---                                                          
#> 55996:    996 2020-04-22      14 3.873626e+06 case      ensemble
#> 55997:    997 2020-04-22      14 2.679323e+06 case      ensemble
#> 55998:    998 2020-04-22      14 1.221193e+06 case      ensemble
#> 55999:    999 2020-04-22      14 1.493394e+06 case      ensemble
#> 56000:   1000 2020-04-22      14 1.214069e+06 case      ensemble
#> 
#> $summarised
#>           date type forecast_type       bottom          top        lower
#>  1: 2020-04-09 case          case 1.137436e+03 4.006069e+06 1.773834e+05
#>  2: 2020-04-10 case          case 5.191559e+03 3.909127e+06 3.330077e+04
#>  3: 2020-04-11 case          case 1.088377e+03 4.077188e+06 5.451655e+05
#>  4: 2020-04-12 case          case 6.706600e+03 4.009126e+06 1.765582e+05
#>  5: 2020-04-13 case          case 1.382070e+04 4.220663e+06 1.118898e+04
#>  6: 2020-04-14 case          case 6.433452e+03 4.095473e+06 1.645612e+05
#>  7: 2020-04-15 case          case 4.932648e+03 4.090818e+06 4.116492e+04
#>  8: 2020-04-16 case          case 2.471047e+04 4.234743e+06 3.768320e+04
#>  9: 2020-04-17 case          case 8.450442e+02 4.170036e+06 1.017453e+05
#> 10: 2020-04-18 case          case 1.004919e+04 4.214568e+06 3.802872e+05
#> 11: 2020-04-19 case          case 2.373637e+04 4.849991e+06 2.476029e+04
#> 12: 2020-04-20 case          case 1.903145e+03 5.053584e+06 1.429957e+05
#> 13: 2020-04-21 case          case 4.934448e+02 5.331966e+06 4.765216e+03
#> 14: 2020-04-22 case          case 1.580793e+04 6.133330e+06 9.710861e+04
#> 15: 2020-04-09 case      ensemble 4.223865e+05 3.410380e+06 7.944490e+05
#> 16: 2020-04-10 case      ensemble 2.079107e+05 3.332013e+06 9.208014e+05
#> 17: 2020-04-11 case      ensemble 4.647461e+05 3.536126e+06 7.649402e+05
#> 18: 2020-04-12 case      ensemble 2.632016e+05 3.377165e+06 9.816192e+05
#> 19: 2020-04-13 case      ensemble 2.962210e+05 3.366492e+06 8.201110e+05
#> 20: 2020-04-14 case      ensemble 2.054910e+05 3.384986e+06 1.014355e+06
#> 21: 2020-04-15 case      ensemble 4.445330e+05 3.551966e+06 7.715118e+05
#> 22: 2020-04-16 case      ensemble 2.984760e+05 3.386283e+06 1.030128e+06
#> 23: 2020-04-17 case      ensemble 1.879403e+05 3.418723e+06 8.661478e+05
#> 24: 2020-04-18 case      ensemble 3.487763e+05 3.561283e+06 1.092577e+06
#> 25: 2020-04-19 case      ensemble 3.908456e+05 3.861011e+06 1.085207e+06
#> 26: 2020-04-20 case      ensemble 3.172326e+05 3.836888e+06 8.015755e+05
#> 27: 2020-04-21 case      ensemble 1.960811e+05 4.094347e+06 1.117499e+06
#> 28: 2020-04-22 case      ensemble 2.012791e+05 4.330901e+06 1.128971e+06
#> 29: 2020-04-09 case            rt 1.877300e+04 3.818077e+06 4.295600e+04
#> 30: 2020-04-10 case            rt 2.747000e+03 3.748166e+06 1.576150e+05
#> 31: 2020-04-11 case            rt 6.330000e+02 3.972324e+06 9.172900e+04
#> 32: 2020-04-12 case            rt 3.308000e+03 3.831723e+06 1.269200e+04
#> 33: 2020-04-13 case            rt 6.056000e+03 3.628530e+06 1.016910e+05
#> 34: 2020-04-14 case            rt 1.050000e+03 3.733141e+06 8.155000e+04
#> 35: 2020-04-15 case            rt 4.820000e+03 3.667840e+06 4.820000e+03
#> 36: 2020-04-16 case            rt 3.290000e+02 3.617539e+06 3.113200e+04
#> 37: 2020-04-17 case            rt 1.398000e+03 3.818905e+06 4.280200e+04
#> 38: 2020-04-18 case            rt 1.595000e+03 3.853004e+06 2.685000e+03
#> 39: 2020-04-19 case            rt 4.520000e+02 3.885130e+06 4.520000e+02
#> 40: 2020-04-20 case            rt 3.147000e+03 3.866614e+06 3.147000e+03
#> 41: 2020-04-21 case            rt 1.000000e+03 3.928455e+06 2.183160e+05
#> 42: 2020-04-22 case            rt 1.107000e+03 3.896843e+06 2.557300e+04
#> 43: 2020-04-09   rt          <NA> 4.751177e-04 1.980707e+00 4.175840e-01
#> 44: 2020-04-10   rt          <NA> 4.682287e-02 2.001927e+00 5.184973e-01
#> 45: 2020-04-11   rt          <NA> 5.086254e-02 1.970387e+00 6.338417e-01
#> 46: 2020-04-12   rt          <NA> 1.000350e-01 2.006976e+00 4.278122e-01
#> 47: 2020-04-13   rt          <NA> 9.094740e-03 1.985910e+00 6.592790e-01
#> 48: 2020-04-14   rt          <NA> 7.537406e-03 1.937348e+00 3.550656e-01
#> 49: 2020-04-15   rt          <NA> 1.564573e-01 2.079000e+00 4.519936e-01
#> 50: 2020-04-16   rt          <NA> 3.868733e-03 1.990346e+00 4.782888e-01
#> 51: 2020-04-17   rt          <NA> 1.021447e-02 1.972224e+00 5.092574e-01
#> 52: 2020-04-18   rt          <NA> 5.283051e-03 2.031273e+00 5.969033e-01
#> 53: 2020-04-19   rt          <NA> 1.097677e-01 2.072376e+00 4.675617e-01
#> 54: 2020-04-20   rt          <NA> 8.387041e-02 2.033292e+00 4.153957e-01
#> 55: 2020-04-21   rt          <NA> 2.291354e-03 2.042227e+00 4.226281e-01
#> 56: 2020-04-22   rt          <NA> 1.078902e-01 2.134185e+00 5.143311e-01
#>           date type forecast_type       bottom          top        lower
#>            upper       median         mean           sd
#>  1: 1.772067e+06 1.620199e+06 1.945822e+06 1.431245e+06
#>  2: 1.526606e+06 1.520140e+06 1.861226e+06 1.457551e+06
#>  3: 2.248827e+06 1.800747e+06 2.013418e+06 1.433271e+06
#>  4: 1.839157e+06 1.685617e+06 1.975138e+06 1.446599e+06
#>  5: 1.723921e+06 1.718501e+06 2.026876e+06 1.481479e+06
#>  6: 1.819907e+06 1.689477e+06 1.983566e+06 1.494320e+06
#>  7: 1.795928e+06 1.766716e+06 2.042720e+06 1.521658e+06
#>  8: 1.739150e+06 1.725680e+06 2.046613e+06 1.500864e+06
#>  9: 1.853307e+06 1.753599e+06 2.036039e+06 1.504792e+06
#> 10: 2.234972e+06 1.948522e+06 2.158359e+06 1.485340e+06
#> 11: 2.025962e+06 2.020071e+06 2.358345e+06 1.701693e+06
#> 12: 2.155906e+06 2.043732e+06 2.404931e+06 1.841474e+06
#> 13: 2.081033e+06 2.078207e+06 2.577156e+06 2.061721e+06
#> 14: 2.475370e+06 2.406683e+06 2.944792e+06 2.365352e+06
#> 15: 2.103337e+06 1.792470e+06 1.901061e+06 9.974247e+05
#> 16: 2.237571e+06 1.743558e+06 1.865980e+06 9.942945e+05
#> 17: 2.085872e+06 1.800930e+06 1.949582e+06 1.000719e+06
#> 18: 2.347514e+06 1.785327e+06 1.914671e+06 1.024951e+06
#> 19: 2.157792e+06 1.806574e+06 1.901381e+06 9.977117e+05
#> 20: 2.316299e+06 1.812989e+06 1.923528e+06 1.014460e+06
#> 21: 2.081266e+06 1.761738e+06 1.901794e+06 9.933646e+05
#> 22: 2.330931e+06 1.755278e+06 1.916964e+06 1.027973e+06
#> 23: 2.253899e+06 1.814037e+06 1.941751e+06 1.059268e+06
#> 24: 2.371800e+06 1.848731e+06 1.988012e+06 1.036146e+06
#> 25: 2.503554e+06 1.933163e+06 2.113046e+06 1.123609e+06
#> 26: 2.318089e+06 2.031159e+06 2.142156e+06 1.156001e+06
#> 27: 2.617410e+06 2.011215e+06 2.247883e+06 1.266519e+06
#> 28: 2.813815e+06 2.221485e+06 2.427234e+06 1.400557e+06
#> 29: 1.596867e+06 1.571623e+06 1.856300e+06 1.405201e+06
#> 30: 1.723316e+06 1.589956e+06 1.870735e+06 1.361683e+06
#> 31: 1.634593e+06 1.573039e+06 1.885746e+06 1.432933e+06
#> 32: 1.542029e+06 1.530770e+06 1.854203e+06 1.399248e+06
#> 33: 1.606865e+06 1.511076e+06 1.775886e+06 1.364671e+06
#> 34: 1.637758e+06 1.589264e+06 1.863490e+06 1.379290e+06
#> 35: 1.474479e+06 1.473958e+06 1.760869e+06 1.353464e+06
#> 36: 1.513568e+06 1.485584e+06 1.787314e+06 1.394301e+06
#> 37: 1.649232e+06 1.609652e+06 1.847463e+06 1.417252e+06
#> 38: 1.550703e+06 1.546552e+06 1.817664e+06 1.424194e+06
#> 39: 1.662103e+06 1.660460e+06 1.867747e+06 1.399447e+06
#> 40: 1.608471e+06 1.607664e+06 1.879381e+06 1.431287e+06
#> 41: 1.832943e+06 1.657627e+06 1.918611e+06 1.413487e+06
#> 42: 1.606586e+06 1.595559e+06 1.909677e+06 1.436664e+06
#> 43: 1.289713e+00 1.025020e+00 1.096687e+00 6.317528e-01
#> 44: 1.412843e+00 1.057046e+00 1.104377e+00 6.474701e-01
#> 45: 1.472332e+00 1.043022e+00 1.092441e+00 6.208606e-01
#> 46: 1.332215e+00 1.070326e+00 1.114393e+00 6.200150e-01
#> 47: 1.553963e+00 1.062507e+00 1.113331e+00 6.425795e-01
#> 48: 1.272011e+00 1.012147e+00 1.091127e+00 6.567634e-01
#> 49: 1.329778e+00 1.112736e+00 1.129766e+00 6.162113e-01
#> 50: 1.371395e+00 1.078113e+00 1.125621e+00 6.440276e-01
#> 51: 1.410291e+00 1.036061e+00 1.097205e+00 6.333096e-01
#> 52: 1.440493e+00 1.082258e+00 1.130660e+00 6.403728e-01
#> 53: 1.360729e+00 1.095507e+00 1.126867e+00 6.344440e-01
#> 54: 1.288352e+00 1.041814e+00 1.105942e+00 6.177932e-01
#> 55: 1.386322e+00 1.061226e+00 1.115434e+00 6.677686e-01
#> 56: 1.409120e+00 1.078006e+00 1.147321e+00 6.539430e-01
#>            upper       median         mean           sd
#> 

# }