Forecast Infections and the Time-Varying Reproduction Number

This function provides optional tools for forecasting cases and Rt estimates using the timeseries methods (via the EpiSoon package). It requires the Episoon package. Installation instructions for the EpiSoon package are available here.

forecast_infections(
  infections,
  rts,
  gt_mean,
  gt_sd,
  gt_max = 30,
  ensemble_type = "mean",
  forecast_model,
  CrIs = c(0.2, 0.5, 0.9),
  horizon = 14,
  samples = 1000
)

Arguments

infections
A data frame of cases by date of infection containing the following variables: date, mean, sd

rts
A data frame of Rt estimates by date of infection containing the following variables: date, mean, sd

gt_mean
Numeric, the mean of the gamma distributed generation time.

gt_sd
Numeric, the standard deviation of the gamma distributed generation time.

gt_max
Numeric, the maximum allowed value of the gamma distributed generation time.

ensemble_type
Character string indicating the type of ensemble to use. By default this is an unweighted ensemble ("mean") with no other types currently supported.

forecast_model
An uninitialised forecast model function to be passed to EpiSoon::forecast_rt. Used for forecasting future Rt and case co An example of the required structure is: function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}.

CrIs
Numeric vector of credible intervals to calculate.

horizon
Numeric, defaults to 14. The horizon over which to forecast Rts and cases.

samples
Numeric, the number of forecast samples to take.

Value

A list of data.tables. The first entry ("samples") contains raw forecast samples and the second entry ("summarised") contains summarised forecasts.

Examples

# \donttest{
if(requireNamespace("EpiSoon")){
   if(requireNamespace("forecastHybrid")){
# example case data
reported_cases <- EpiNow2::example_confirmed[1:40]
 
generation_time <- get_generation_time(disease = "SARS-CoV-2", source = "ganyani")
incubation_period <- get_incubation_period(disease = "SARS-CoV-2", source = "lauer")
reporting_delay <- EpiNow2::bootstrapped_dist_fit(rlnorm(100, log(6), 1), max_value = 15)

# estimate Rt and infections from data
out <- estimate_infections(reported_cases, generation_time = generation_time,
                           delays = list(incubation_period, reporting_delay),
                           gp = list(), samples = 100,
                           stan_args = list(cores = ifelse(interactive(), 4, 1), 
                                            warmup = 100, chains = 4))

# forecast Rt and infections from estimates
forecast <- forecast_infections(
    infections = out$summarised[variable == "infections"],
    rts = out$summarised[variable == "R"],
     gt_mean = out$summarised[variable == "gt_mean"]$mean,
     gt_sd = out$summarised[variable == "gt_sd"]$mean,
     gt_max = 30,
     forecast_model = function(y, ...){
       EpiSoon::forecastHybrid_model(y = y[max(1, length(y) - 21):length(y)],
       model_params = list(models = "aefz", weights = "equal"),
       forecast_params = list(PI.combination = "mean"), ...)},
     horizon = 14, 
     samples = 1000)
                                
forecast$summarised
  }
 }

#> Loading required namespace: EpiSoon

#> Loading required namespace: forecastHybrid

#> Registered S3 method overwritten by 'quantmod':
#>   method            from
#>   as.zoo.data.frame zoo 

#>           date type forecast_type       median         mean           sd
#>  1: 2020-04-09 case          case 4.085348e+04 47685.362704 3.566216e+04
#>  2: 2020-04-10 case          case 4.170101e+04 46994.322673 3.356199e+04
#>  3: 2020-04-11 case          case 4.038192e+04 46468.680402 3.405321e+04
#>  4: 2020-04-12 case          case 4.493122e+04 50188.239795 3.498278e+04
#>  5: 2020-04-13 case          case 4.661101e+04 51963.361312 3.638459e+04
#>  6: 2020-04-14 case          case 4.295913e+04 50042.140061 3.555027e+04
#>  7: 2020-04-15 case          case 4.834441e+04 53718.351054 3.649525e+04
#>  8: 2020-04-16 case          case 4.696925e+04 53304.082996 3.797803e+04
#>  9: 2020-04-17 case          case 4.807039e+04 53901.236629 3.644408e+04
#> 10: 2020-04-18 case          case 4.885875e+04 55272.874712 3.790209e+04
#> 11: 2020-04-19 case          case 4.951305e+04 55884.684967 3.855769e+04
#> 12: 2020-04-20 case          case 4.874365e+04 54630.914859 3.852186e+04
#> 13: 2020-04-21 case          case 4.905868e+04 55451.103158 3.819649e+04
#> 14: 2020-04-22 case          case 5.278885e+04 59751.286734 4.141114e+04
#> 15: 2020-04-09 case      ensemble 4.270909e+04 45765.505352 2.421275e+04
#> 16: 2020-04-10 case      ensemble 4.280798e+04 45659.626836 2.388352e+04
#> 17: 2020-04-11 case      ensemble 4.234147e+04 45552.622201 2.398442e+04
#> 18: 2020-04-12 case      ensemble 4.475448e+04 47341.114898 2.359890e+04
#> 19: 2020-04-13 case      ensemble 4.480587e+04 48055.817656 2.503236e+04
#> 20: 2020-04-14 case      ensemble 4.447154e+04 47668.099530 2.438108e+04
#> 21: 2020-04-15 case      ensemble 4.586556e+04 49035.180027 2.473179e+04
#> 22: 2020-04-16 case      ensemble 4.604161e+04 49562.056998 2.573009e+04
#> 23: 2020-04-17 case      ensemble 4.708635e+04 49289.205815 2.440832e+04
#> 24: 2020-04-18 case      ensemble 4.679220e+04 50228.968356 2.583411e+04
#> 25: 2020-04-19 case      ensemble 4.905275e+04 50699.760984 2.558120e+04
#> 26: 2020-04-20 case      ensemble 4.601081e+04 49174.324430 2.454644e+04
#> 27: 2020-04-21 case      ensemble 4.647834e+04 49941.947079 2.654122e+04
#> 28: 2020-04-22 case      ensemble 5.046899e+04 52769.383867 2.591248e+04
#> 29: 2020-04-09 case            rt 3.686000e+04 43845.648000 3.354823e+04
#> 30: 2020-04-10 case            rt 3.691950e+04 44324.931000 3.320120e+04
#> 31: 2020-04-11 case            rt 3.760800e+04 44636.564000 3.384796e+04
#> 32: 2020-04-12 case            rt 3.839550e+04 44493.990000 3.264347e+04
#> 33: 2020-04-13 case            rt 3.791050e+04 44148.274000 3.305241e+04
#> 34: 2020-04-14 case            rt 3.862550e+04 45294.059000 3.301212e+04
#> 35: 2020-04-15 case            rt 3.785700e+04 44352.009000 3.247234e+04
#> 36: 2020-04-16 case            rt 3.865450e+04 45820.031000 3.441177e+04
#> 37: 2020-04-17 case            rt 3.849800e+04 44677.175000 3.396924e+04
#> 38: 2020-04-18 case            rt 3.890000e+04 45185.062000 3.398775e+04
#> 39: 2020-04-19 case            rt 3.856200e+04 45514.837000 3.305841e+04
#> 40: 2020-04-20 case            rt 3.766100e+04 43717.734000 3.224941e+04
#> 41: 2020-04-21 case            rt 3.801100e+04 44432.791000 3.327585e+04
#> 42: 2020-04-22 case            rt 3.822200e+04 45787.481000 3.405489e+04
#> 43: 2020-04-09   rt          <NA> 9.362423e-01     1.067070 6.927807e-01
#> 44: 2020-04-10   rt          <NA> 9.681498e-01     1.068322 6.707448e-01
#> 45: 2020-04-11   rt          <NA> 1.000556e+00     1.067576 6.612737e-01
#> 46: 2020-04-12   rt          <NA> 1.078905e+00     1.129257 6.561683e-01
#> 47: 2020-04-13   rt          <NA> 1.082763e+00     1.137645 6.996854e-01
#> 48: 2020-04-14   rt          <NA> 1.032099e+00     1.122312 6.807788e-01
#> 49: 2020-04-15   rt          <NA> 1.056724e+00     1.129633 7.083597e-01
#> 50: 2020-04-16   rt          <NA> 1.092685e+00     1.165610 7.021801e-01
#> 51: 2020-04-17   rt          <NA> 1.174301e+00     1.217323 7.090532e-01
#> 52: 2020-04-18   rt          <NA> 1.107157e+00     1.166704 7.060898e-01
#> 53: 2020-04-19   rt          <NA> 1.134159e+00     1.225982 7.061670e-01
#> 54: 2020-04-20   rt          <NA> 1.136757e+00     1.207617 7.143692e-01
#> 55: 2020-04-21   rt          <NA> 1.171346e+00     1.223800 7.204862e-01
#> 56: 2020-04-22   rt          <NA> 1.195450e+00     1.224273 6.889118e-01
#>           date type forecast_type       median         mean           sd
#>         lower_90     lower_50     lower_20     upper_20     upper_50
#>  1: 4.211321e+03 2.027275e+04 3.235109e+04 49991.342842 65763.529446
#>  2: 4.196443e+03 2.121257e+04 3.210000e+04 50653.703829 67489.686281
#>  3: 4.081872e+03 1.983072e+04 3.093273e+04 50672.622306 65904.352449
#>  4: 5.143139e+03 2.211278e+04 3.534926e+04 53457.681899 70430.417894
#>  5: 4.862313e+03 2.304315e+04 3.566530e+04 55116.496681 73850.482157
#>  6: 5.638493e+03 2.243797e+04 3.470372e+04 52500.176859 72259.949233
#>  7: 6.463673e+03 2.406978e+04 3.872615e+04 58043.966912 77105.472802
#>  8: 4.832394e+03 2.203234e+04 3.780153e+04 58999.573461 75393.927710
#>  9: 5.390453e+03 2.445309e+04 3.938843e+04 59733.654166 76144.769439
#> 10: 5.623175e+03 2.539595e+04 3.863801e+04 60853.437114 77116.529277
#> 11: 5.292826e+03 2.511655e+04 3.914261e+04 61739.400967 81477.840268
#> 12: 4.773045e+03 2.422623e+04 3.845023e+04 59116.862868 77844.614638
#> 13: 5.976300e+03 2.522196e+04 3.890076e+04 60133.382541 79192.561570
#> 14: 6.602983e+03 2.656382e+04 4.193430e+04 66056.163079 84119.303102
#> 15: 1.138070e+04 2.843673e+04 3.624461e+04 48268.282964 60412.445147
#> 16: 1.247965e+04 2.786423e+04 3.698794e+04 49151.280499 60556.695592
#> 17: 1.115696e+04 2.789355e+04 3.720893e+04 48861.736191 59609.557393
#> 18: 1.345071e+04 2.939819e+04 3.845903e+04 50546.936538 63598.531706
#> 19: 1.300662e+04 2.897124e+04 3.907067e+04 52137.823891 63540.924320
#> 20: 1.366956e+04 2.856507e+04 3.923820e+04 51819.478333 62535.123522
#> 21: 1.335954e+04 3.100614e+04 4.005016e+04 52019.360101 65564.210788
#> 22: 1.292062e+04 3.069143e+04 4.021958e+04 52417.183689 65938.518276
#> 23: 1.407433e+04 3.127937e+04 4.011381e+04 53592.419925 63667.809608
#> 24: 1.374987e+04 3.085192e+04 4.007838e+04 53651.526637 65599.902683
#> 25: 1.428895e+04 3.128312e+04 4.116845e+04 55730.437311 66013.129982
#> 26: 1.353597e+04 3.047876e+04 3.992572e+04 53353.058725 65484.477424
#> 27: 1.272838e+04 2.981696e+04 3.964810e+04 53882.519897 65722.520355
#> 28: 1.520186e+04 3.432524e+04 4.451273e+04 57103.511287 68141.923705
#> 29: 3.660950e+03 1.729950e+04 2.816920e+04 45779.600000 64248.250000
#> 30: 3.678600e+03 1.879675e+04 2.976360e+04 44880.200000 63524.500000
#> 31: 3.477650e+03 1.757325e+04 2.907280e+04 47414.200000 65615.500000
#> 32: 3.011650e+03 1.906025e+04 3.016700e+04 47416.400000 62815.500000
#> 33: 3.048050e+03 1.844450e+04 2.764840e+04 47445.200000 63633.500000
#> 34: 4.624600e+03 1.862350e+04 3.084820e+04 48913.600000 65211.750000
#> 35: 4.312400e+03 1.895000e+04 3.044920e+04 47598.600000 62070.250000
#> 36: 3.140350e+03 1.913200e+04 2.956000e+04 49553.600000 67321.500000
#> 37: 2.473700e+03 1.723675e+04 2.892580e+04 47195.400000 63738.750000
#> 38: 2.978650e+03 1.838950e+04 3.031200e+04 48723.600000 64404.750000
#> 39: 3.806750e+03 1.895175e+04 3.069760e+04 48870.800000 65432.500000
#> 40: 3.041900e+03 1.878675e+04 2.900260e+04 46643.000000 63962.750000
#> 41: 3.488100e+03 1.836500e+04 3.005080e+04 46925.400000 63616.500000
#> 42: 4.326950e+03 1.763775e+04 2.942400e+04 48848.200000 65999.500000
#> 43: 1.148270e-01 5.335565e-01 7.803615e-01     1.178532     1.548991
#> 44: 1.226617e-01 5.455878e-01 8.026072e-01     1.178914     1.509701
#> 45: 1.408044e-01 5.336658e-01 8.287537e-01     1.196758     1.506476
#> 46: 1.753514e-01 6.176444e-01 8.925810e-01     1.249398     1.544244
#> 47: 1.286633e-01 5.712464e-01 8.942993e-01     1.278262     1.593773
#> 48: 1.686476e-01 5.900834e-01 8.666938e-01     1.250237     1.582463
#> 49: 1.369448e-01 5.555331e-01 8.693106e-01     1.263673     1.576928
#> 50: 1.826676e-01 6.009410e-01 8.809755e-01     1.319027     1.638352
#> 51: 1.378412e-01 6.438948e-01 9.660149e-01     1.368464     1.718966
#> 52: 1.248296e-01 6.064430e-01 9.122986e-01     1.304352     1.655485
#> 53: 1.697079e-01 6.868219e-01 9.733983e-01     1.356738     1.676416
#> 54: 1.686103e-01 6.539484e-01 9.415215e-01     1.303321     1.681817
#> 55: 1.739196e-01 6.657569e-01 9.724750e-01     1.379087     1.718782
#> 56: 1.809191e-01 7.129266e-01 9.830946e-01     1.391035     1.677858
#>         lower_90     lower_50     lower_20     upper_20     upper_50
#>         upper_90
#>  1: 1.192798e+05
#>  2: 1.104054e+05
#>  3: 1.102968e+05
#>  4: 1.170790e+05
#>  5: 1.207574e+05
#>  6: 1.171907e+05
#>  7: 1.214862e+05
#>  8: 1.255139e+05
#>  9: 1.204518e+05
#> 10: 1.257605e+05
#> 11: 1.262294e+05
#> 12: 1.267231e+05
#> 13: 1.248160e+05
#> 14: 1.376374e+05
#> 15: 9.202266e+04
#> 16: 8.781274e+04
#> 17: 9.126366e+04
#> 18: 8.982480e+04
#> 19: 9.342482e+04
#> 20: 9.353576e+04
#> 21: 9.305571e+04
#> 22: 9.475081e+04
#> 23: 9.304560e+04
#> 24: 9.744766e+04
#> 25: 9.697544e+04
#> 26: 9.405805e+04
#> 27: 9.999713e+04
#> 28: 1.003857e+05
#> 29: 1.062892e+05
#> 30: 1.104006e+05
#> 31: 1.065996e+05
#> 32: 1.098644e+05
#> 33: 1.070319e+05
#> 34: 1.081132e+05
#> 35: 1.087040e+05
#> 36: 1.135331e+05
#> 37: 1.115333e+05
#> 38: 1.088660e+05
#> 39: 1.095931e+05
#> 40: 1.026895e+05
#> 41: 1.083854e+05
#> 42: 1.106921e+05
#> 43: 2.298320e+00
#> 44: 2.321865e+00
#> 45: 2.235196e+00
#> 46: 2.336059e+00
#> 47: 2.388483e+00
#> 48: 2.337572e+00
#> 49: 2.436588e+00
#> 50: 2.438534e+00
#> 51: 2.452093e+00
#> 52: 2.439577e+00
#> 53: 2.469280e+00
#> 54: 2.515757e+00
#> 55: 2.484731e+00
#> 56: 2.414579e+00
#>         upper_90

# }