Get Combined Regional Results

Get Combined Regional Results

get_regional_results(results_dir, date, forecast = FALSE)

Arguments

results_dir
A character string indicating the folder containing the EpiNow2 results to extract.

date
A Character string (in the format "yyyy-mm-dd") indicating the date to extract data for. Defaults to "latest" which finds the latest results available.

forecast
Logical, defaults to FALSE. Should forecast results be returned.

Value

A list of estimates, forecasts and estimated cases by date of report.

Examples

if (FALSE) {
# see ?regional_summary for code to produce test results
out <- get_regional_results("../test", forecast = TRUE)

out
}
## Code
get_regional_results

#> function (results_dir, date, forecast = FALSE) 
#> {
#>     if (missing(date)) {
#>         date <- "latest"
#>     }
#>     regions <- list.files(results_dir, recursive = FALSE)
#>     names(regions) <- regions
#>     load_data <- purrr::safely(EpiNow2::get_raw_result)
#>     get_estimates <- function(samples, summarised) {
#>         samples <- purrr::map(regions, ~load_data(samples, ., 
#>             result_dir = results_dir, date = date)[[1]])
#>         samples <- data.table::rbindlist(samples, idcol = "region")
#>         summarised <- purrr::map(regions, ~load_data(summarised, 
#>             ., result_dir = results_dir, date = date)[[1]])
#>         summarised <- data.table::rbindlist(summarised, idcol = "region")
#>         out <- list()
#>         out$samples <- samples
#>         out$summarised <- summarised
#>         return(out)
#>     }
#>     out <- list()
#>     out$estimates <- get_estimates(samples = "estimate_samples.rds", 
#>         summarised = "summarised_estimates.rds")
#>     if (forecast) {
#>         out$forecast <- get_estimates(samples = "forecast_samples.rds", 
#>             summarised = "summarised_forecast.rds")
#>         out$estimated_reported_cases <- get_estimates(samples = "estimated_reported_cases_samples.rds", 
#>             summarised = "summarised_estimated_reported_cases.rds")
#>     }
#>     return(out)
#> }
#> <bytecode: 0x5590c5202818>
#> <environment: namespace:EpiNow2>