Real-time Rt Estimation, Forecasting and Reporting by Region

Estimates Rt by region. See the documentation for epinow for further information.

regional_epinow(
  reported_cases,
  target_folder,
  target_date,
  case_limit = 20,
  cores = 1,
  return_estimates = TRUE,
  ...
)

Arguments

reported_cases
A data frame of confirmed cases (confirm) by date (date), and region (region).

target_folder
Character string specifying where to save results (will create if not present).

target_date
Date, defaults to maximum found in the data if not specified.

case_limit
Numeric, the minimum number of cases in a region required for that region to be evaluated. Defaults to 20.

cores
Numeric, defaults to 2. The number of cores to use when fitting the stan model.

return_estimates
Logical, defaults to TRUE. Should estimates be returned.

...
Pass additional arguments to epinow

Value

Examples

 if (FALSE) {
## Construct example distributions
generation_time <- list(mean = EpiNow2::covid_generation_times[1, ]$mean,
                        mean_sd = EpiNow2::covid_generation_times[1, ]$mean_sd,
                        sd = EpiNow2::covid_generation_times[1, ]$sd,
                        sd_sd = EpiNow2::covid_generation_times[1, ]$sd_sd,
                        max = 30)

incubation_period <- list(mean = EpiNow2::covid_incubation_period[1, ]$mean,
                          mean_sd = EpiNow2::covid_incubation_period[1, ]$mean_sd,
                          sd = EpiNow2::covid_incubation_period[1, ]$sd,
                          sd_sd = EpiNow2::covid_incubation_period[1, ]$sd_sd,
                          max = 30)

reporting_delay <- list(mean = log(10),
                        mean_sd = 0.8,
                        sd = log(2),
                        sd_sd = 0.1,
                        max = 30)

## Uses example case vector
cases <- EpiNow2::example_confirmed

cases <- data.table::rbindlist(list(
  data.table::copy(cases)[, region := "testland"],
  cases[, region := "realland"]))

## Run basic nowcasting pipeline
out <- regional_epinow(reported_cases = cases,
                       generation_time = generation_time,
                       incubation_period = incubation_period,
                       reporting_delay = reporting_delay,
                       samples = 1000, warmup = 500, cores = 2, chains = 2,
                       verbose = TRUE)
}