Real-time Rt Estimation, Forecasting and Reporting by Region

Estimates Rt by region. See the documentation for epinow for further information. The progress of producing estimates across multiple regions is tracked using the progressr package. Modify this behaviour using progressr::handlers and enable it in batch by setting R_PROGRESSR_ENABLE=TRUE as an environment variable.

regional_epinow(
  reported_cases,
  target_folder = NULL,
  target_date,
  non_zero_points = 2,
  output = c("regions", "summary", "samples", "plots", "latest"),
  return_output = FALSE,
  summary_args = list(),
  logs = tempdir(),
  ...
)

Arguments

reported_cases
A data frame of confirmed cases (confirm) by date (date), and region (region).

target_folder
Character string specifying where to save results (will create if not present).

target_date
Date, defaults to maximum found in the data if not specified.

non_zero_points
Numeric, the minimum number of time points with non-zero cases in a region required for that region to be evaluated. Defaults to 2.

output
A character vector of optional output to return. Supported options are the individual regional estimates ("regions"), samples ("samples"), plots ("plots"), copying the individual region dated folder into a latest folder (if target_folder is not null - set using "latest"), the stan fit of the underlying model ("fit"), and an overall summary across regions ("summary"). The default is to return samples and plots alongside summarised estimates and summary statistics. If target_folder is not NULL then the default is also to copy all results into a latest folder.

return_output
Logical, defaults to FALSE. Should output be returned, this automatically updates to TRUE if no directory for saving is specified.

summary_args
A list of arguments passed to regional_summary. See the regional_summary documentation for details.

logs
Character path indicating the target folder in which to store log information. Defaults to the temporary directory if not specified. Default logging can be disabled if logs is set to NULL. If specifying a custom logging setup then the code for setup_default_logging and the setup_logging function are a sensible place to start.

...
Pass additional arguments to epinow. See the documentation for epinow for details.

Value

A list of output stratified at the top level into regional output and across region output summary output

Examples

 # \donttest{
# construct example distributions
generation_time <- get_generation_time(disease = "SARS-CoV-2", source = "ganyani")
incubation_period <- get_incubation_period(disease = "SARS-CoV-2", source = "lauer")
reporting_delay <- list(mean = log(10), mean_sd = log(2),
                        sd = log(2), sd_sd = log(1.1), max = 30)
                        
# uses example case vector
cases <- EpiNow2::example_confirmed[1:40]
cases <- data.table::rbindlist(list(
  data.table::copy(cases)[, region := "testland"],
  cases[, region := "realland"]))
                
# run epinow across multiple regions and generate summaries
# samples and warmup have been reduced for this example
out <- regional_epinow(reported_cases = cases, 
                       samples = 100,
                       generation_time = generation_time,
                       delays = list(incubation_period, reporting_delay),
                       stan_args = list(warmup = 100, 
                                        cores = ifelse(interactive(), 4, 1)))

#> INFO [2020-10-17 11:34:34] Producing following optional outputs: regions, summary, samples, plots, latest

#> Logging threshold set at INFO for the EpiNow2 logger

#> Writing EpiNow2 logs to: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpDy5hT0/regional-epinow/2020-04-01.log

#> Logging threshold set at INFO for the EpiNow2.epinow logger

#> Writing EpiNow2.epinow logs to: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpDy5hT0/epinow/2020-04-01.log

#> Logging threshold set at INFO for the EpiNow2.epinow.estimate_infections logger

#> Writing EpiNow2.epinow.estimate_infections logs to: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpDy5hT0/epinow-estimate-infections/2020-04-01.log

#> INFO [2020-10-17 11:34:34] Reporting estimates using data up to: 2020-04-01
#> INFO [2020-10-17 11:34:34] No target directory specified so returning output
#> INFO [2020-10-17 11:34:34] Producing estimates for: testland, realland
#> INFO [2020-10-17 11:34:34] Regions excluded: none
#> INFO [2020-10-17 11:34:34] Showing progress using progressr. Modify this behaviour using progressr::handlers.
#> INFO [2020-10-17 11:37:17] Completed estimates for: testland
#> INFO [2020-10-17 11:40:09] Completed estimates for: realland
#> INFO [2020-10-17 11:40:09] Completed regional estimates
#> INFO [2020-10-17 11:40:09] Regions with estimates: 2
#> INFO [2020-10-17 11:40:09] Regions with runtime errors: 0
#> INFO [2020-10-17 11:40:09] Producing summary
#> INFO [2020-10-17 11:40:09] No summary directory specified so returning summary output
#> INFO [2020-10-17 11:40:09] No target directory specified so returning timings

# }