Real-time Rt Estimation, Forecasting and Reporting by Region

Estimates Rt by region. See the documentation for epinow for further information.

regional_epinow(
  reported_cases,
  target_folder,
  target_date,
  non_zero_points = 2,
  cores = 1,
  summary = TRUE,
  summary_dir,
  region_scale = "Region",
  all_regions_summary = TRUE,
  return_estimates = TRUE,
  max_plot = 10,
  ...
)

Arguments

reported_cases
A data frame of confirmed cases (confirm) by date (date), and region (region).

target_folder
Character string specifying where to save results (will create if not present).

target_date
Date, defaults to maximum found in the data if not specified.

non_zero_points
Numeric, the minimum number of time points with non-zero cases in a region required for that region to be evaluated. Defaults to 2.

cores
Numeric, defaults to 2. The number of cores to use when fitting the stan model.

summary
Logical, should summary measures be calculated.

summary_dir
A character string giving the directory in which to store summary of results.

region_scale
A character string indicating the name to give the regions being summarised.

all_regions_summary
Logical, defaults to TRUE. Should summary plots for all regions be returned rather than just regions of interest.

return_estimates
Logical, defaults to TRUE. Should estimates be returned.

max_plot
Numeric, defaults to 10. A multiplicative upper bound on the number of cases shown on the plot. Based on the maximum number of reported cases.

...
Pass additional arguments to epinow

Value

A list of output stratified at the top level into regional output and across region output summary output

Examples

 # \donttest{
## Construct example distributions
generation_time <- list(mean = EpiNow2::covid_generation_times[1, ]$mean,
                        mean_sd = EpiNow2::covid_generation_times[1, ]$mean_sd,
                        sd = EpiNow2::covid_generation_times[1, ]$sd,
                        sd_sd = EpiNow2::covid_generation_times[1, ]$sd_sd,
                        max = 30)
                          
incubation_period <- list(mean = EpiNow2::covid_incubation_period[1, ]$mean,
                          mean_sd = EpiNow2::covid_incubation_period[1, ]$mean_sd,
                          sd = EpiNow2::covid_incubation_period[1, ]$sd,
                          sd_sd = EpiNow2::covid_incubation_period[1, ]$sd_sd,
                          max = 30)
                   
reporting_delay <- list(mean = log(10),
                        mean_sd = log(2),
                        sd = log(2),
                        sd_sd = log(1.1),
                        max = 30)
                        
## Uses example case vector
cases <- EpiNow2::example_confirmed[1:40]

cases <- data.table::rbindlist(list(
  data.table::copy(cases)[, region := "testland"],
  cases[, region := "realland"]))
  
## Run basic nowcasting pipeline
## Here we reduce the accuracy of the GP approximation in order to reduce runtime
out <- regional_epinow(reported_cases = cases,
                       generation_time = generation_time,
                       delays = list(incubation_period, reporting_delay),
                       adapt_delta = 0.9,
                       samples = 2000, warmup = 200,
                       cores = ifelse(interactive(), 4, 1), chains = 4)

#> INFO [2020-08-19 12:25:56] Reporting estimates using data up to: 2020-04-01
#> INFO [2020-08-19 12:25:56] Producing estimates for: testland, realland
#> INFO [2020-08-19 12:25:56] Initialising estimates for: testland
#> DEBUG [2020-08-19 12:25:56] Running for 2000 samples (across 4 chains each with a warm up of 200 iterations each) and 65 time steps of which 7 are a forecast
#> 
#> SAMPLING FOR MODEL 'estimate_infections' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000342 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 3.42 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 700 [  0%]  (Warmup)
#> Chain 1: Iteration:  50 / 700 [  7%]  (Warmup)
#> Chain 1: Iteration: 100 / 700 [ 14%]  (Warmup)
#> Chain 1: Iteration: 150 / 700 [ 21%]  (Warmup)
#> Chain 1: Iteration: 200 / 700 [ 28%]  (Warmup)
#> Chain 1: Iteration: 201 / 700 [ 28%]  (Sampling)
#> Chain 1: Iteration: 250 / 700 [ 35%]  (Sampling)
#> Chain 1: Iteration: 300 / 700 [ 42%]  (Sampling)
#> Chain 1: Iteration: 350 / 700 [ 50%]  (Sampling)
#> Chain 1: Iteration: 400 / 700 [ 57%]  (Sampling)
#> Chain 1: Iteration: 450 / 700 [ 64%]  (Sampling)
#> Chain 1: Iteration: 500 / 700 [ 71%]  (Sampling)
#> Chain 1: Iteration: 550 / 700 [ 78%]  (Sampling)
#> Chain 1: Iteration: 600 / 700 [ 85%]  (Sampling)
#> Chain 1: Iteration: 650 / 700 [ 92%]  (Sampling)
#> Chain 1: Iteration: 700 / 700 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 31.7383 seconds (Warm-up)
#> Chain 1:                35.0083 seconds (Sampling)
#> Chain 1:                66.7466 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'estimate_infections' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000296 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 2.96 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:   1 / 700 [  0%]  (Warmup)
#> Chain 2: Iteration:  50 / 700 [  7%]  (Warmup)
#> Chain 2: Iteration: 100 / 700 [ 14%]  (Warmup)
#> Chain 2: Iteration: 150 / 700 [ 21%]  (Warmup)
#> Chain 2: Iteration: 200 / 700 [ 28%]  (Warmup)
#> Chain 2: Iteration: 201 / 700 [ 28%]  (Sampling)
#> Chain 2: Iteration: 250 / 700 [ 35%]  (Sampling)
#> Chain 2: Iteration: 300 / 700 [ 42%]  (Sampling)
#> Chain 2: Iteration: 350 / 700 [ 50%]  (Sampling)
#> Chain 2: Iteration: 400 / 700 [ 57%]  (Sampling)
#> Chain 2: Iteration: 450 / 700 [ 64%]  (Sampling)
#> Chain 2: Iteration: 500 / 700 [ 71%]  (Sampling)
#> Chain 2: Iteration: 550 / 700 [ 78%]  (Sampling)
#> Chain 2: Iteration: 600 / 700 [ 85%]  (Sampling)
#> Chain 2: Iteration: 650 / 700 [ 92%]  (Sampling)
#> Chain 2: Iteration: 700 / 700 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 26.1551 seconds (Warm-up)
#> Chain 2:                33.9962 seconds (Sampling)
#> Chain 2:                60.1513 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'estimate_infections' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000322 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 3.22 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:   1 / 700 [  0%]  (Warmup)
#> Chain 3: Iteration:  50 / 700 [  7%]  (Warmup)
#> Chain 3: Iteration: 100 / 700 [ 14%]  (Warmup)
#> Chain 3: Iteration: 150 / 700 [ 21%]  (Warmup)
#> Chain 3: Iteration: 200 / 700 [ 28%]  (Warmup)
#> Chain 3: Iteration: 201 / 700 [ 28%]  (Sampling)
#> Chain 3: Iteration: 250 / 700 [ 35%]  (Sampling)
#> Chain 3: Iteration: 300 / 700 [ 42%]  (Sampling)
#> Chain 3: Iteration: 350 / 700 [ 50%]  (Sampling)
#> Chain 3: Iteration: 400 / 700 [ 57%]  (Sampling)
#> Chain 3: Iteration: 450 / 700 [ 64%]  (Sampling)
#> Chain 3: Iteration: 500 / 700 [ 71%]  (Sampling)
#> Chain 3: Iteration: 550 / 700 [ 78%]  (Sampling)
#> Chain 3: Iteration: 600 / 700 [ 85%]  (Sampling)
#> Chain 3: Iteration: 650 / 700 [ 92%]  (Sampling)
#> Chain 3: Iteration: 700 / 700 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 27.1209 seconds (Warm-up)
#> Chain 3:                17.6115 seconds (Sampling)
#> Chain 3:                44.7324 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'estimate_infections' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000271 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 2.71 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:   1 / 700 [  0%]  (Warmup)
#> Chain 4: Iteration:  50 / 700 [  7%]  (Warmup)
#> Chain 4: Iteration: 100 / 700 [ 14%]  (Warmup)
#> Chain 4: Iteration: 150 / 700 [ 21%]  (Warmup)
#> Chain 4: Iteration: 200 / 700 [ 28%]  (Warmup)
#> Chain 4: Iteration: 201 / 700 [ 28%]  (Sampling)
#> Chain 4: Iteration: 250 / 700 [ 35%]  (Sampling)
#> Chain 4: Iteration: 300 / 700 [ 42%]  (Sampling)
#> Chain 4: Iteration: 350 / 700 [ 50%]  (Sampling)
#> Chain 4: Iteration: 400 / 700 [ 57%]  (Sampling)
#> Chain 4: Iteration: 450 / 700 [ 64%]  (Sampling)
#> Chain 4: Iteration: 500 / 700 [ 71%]  (Sampling)
#> Chain 4: Iteration: 550 / 700 [ 78%]  (Sampling)
#> Chain 4: Iteration: 600 / 700 [ 85%]  (Sampling)
#> Chain 4: Iteration: 650 / 700 [ 92%]  (Sampling)
#> Chain 4: Iteration: 700 / 700 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 33.3965 seconds (Warm-up)
#> Chain 4:                26.6281 seconds (Sampling)
#> Chain 4:                60.0246 seconds (Total)
#> Chain 4: 
#> INFO [2020-08-19 12:29:50] Completed estimates for: testland

#> Warning: There were 14 divergent transitions after warmup. See
#> http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup
#> to find out why this is a problem and how to eliminate them.

#> Warning: Examine the pairs() plot to diagnose sampling problems

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> INFO [2020-08-19 12:29:50] Initialising estimates for: realland
#> DEBUG [2020-08-19 12:29:50] Running for 2000 samples (across 4 chains each with a warm up of 200 iterations each) and 65 time steps of which 7 are a forecast
#> 
#> SAMPLING FOR MODEL 'estimate_infections' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000356 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 3.56 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 700 [  0%]  (Warmup)
#> Chain 1: Iteration:  50 / 700 [  7%]  (Warmup)
#> Chain 1: Iteration: 100 / 700 [ 14%]  (Warmup)
#> Chain 1: Iteration: 150 / 700 [ 21%]  (Warmup)
#> Chain 1: Iteration: 200 / 700 [ 28%]  (Warmup)
#> Chain 1: Iteration: 201 / 700 [ 28%]  (Sampling)
#> Chain 1: Iteration: 250 / 700 [ 35%]  (Sampling)
#> Chain 1: Iteration: 300 / 700 [ 42%]  (Sampling)
#> Chain 1: Iteration: 350 / 700 [ 50%]  (Sampling)
#> Chain 1: Iteration: 400 / 700 [ 57%]  (Sampling)
#> Chain 1: Iteration: 450 / 700 [ 64%]  (Sampling)
#> Chain 1: Iteration: 500 / 700 [ 71%]  (Sampling)
#> Chain 1: Iteration: 550 / 700 [ 78%]  (Sampling)
#> Chain 1: Iteration: 600 / 700 [ 85%]  (Sampling)
#> Chain 1: Iteration: 650 / 700 [ 92%]  (Sampling)
#> Chain 1: Iteration: 700 / 700 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 33.119 seconds (Warm-up)
#> Chain 1:                36.3418 seconds (Sampling)
#> Chain 1:                69.4608 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'estimate_infections' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000322 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 3.22 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:   1 / 700 [  0%]  (Warmup)
#> Chain 2: Iteration:  50 / 700 [  7%]  (Warmup)
#> Chain 2: Iteration: 100 / 700 [ 14%]  (Warmup)
#> Chain 2: Iteration: 150 / 700 [ 21%]  (Warmup)
#> Chain 2: Iteration: 200 / 700 [ 28%]  (Warmup)
#> Chain 2: Iteration: 201 / 700 [ 28%]  (Sampling)
#> Chain 2: Iteration: 250 / 700 [ 35%]  (Sampling)
#> Chain 2: Iteration: 300 / 700 [ 42%]  (Sampling)
#> Chain 2: Iteration: 350 / 700 [ 50%]  (Sampling)
#> Chain 2: Iteration: 400 / 700 [ 57%]  (Sampling)
#> Chain 2: Iteration: 450 / 700 [ 64%]  (Sampling)
#> Chain 2: Iteration: 500 / 700 [ 71%]  (Sampling)
#> Chain 2: Iteration: 550 / 700 [ 78%]  (Sampling)
#> Chain 2: Iteration: 600 / 700 [ 85%]  (Sampling)
#> Chain 2: Iteration: 650 / 700 [ 92%]  (Sampling)
#> Chain 2: Iteration: 700 / 700 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 29.9601 seconds (Warm-up)
#> Chain 2:                17.7452 seconds (Sampling)
#> Chain 2:                47.7054 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'estimate_infections' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000291 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 2.91 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:   1 / 700 [  0%]  (Warmup)
#> Chain 3: Iteration:  50 / 700 [  7%]  (Warmup)
#> Chain 3: Iteration: 100 / 700 [ 14%]  (Warmup)
#> Chain 3: Iteration: 150 / 700 [ 21%]  (Warmup)
#> Chain 3: Iteration: 200 / 700 [ 28%]  (Warmup)
#> Chain 3: Iteration: 201 / 700 [ 28%]  (Sampling)
#> Chain 3: Iteration: 250 / 700 [ 35%]  (Sampling)
#> Chain 3: Iteration: 300 / 700 [ 42%]  (Sampling)
#> Chain 3: Iteration: 350 / 700 [ 50%]  (Sampling)
#> Chain 3: Iteration: 400 / 700 [ 57%]  (Sampling)
#> Chain 3: Iteration: 450 / 700 [ 64%]  (Sampling)
#> Chain 3: Iteration: 500 / 700 [ 71%]  (Sampling)
#> Chain 3: Iteration: 550 / 700 [ 78%]  (Sampling)
#> Chain 3: Iteration: 600 / 700 [ 85%]  (Sampling)
#> Chain 3: Iteration: 650 / 700 [ 92%]  (Sampling)
#> Chain 3: Iteration: 700 / 700 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 29.7026 seconds (Warm-up)
#> Chain 3:                36.0538 seconds (Sampling)
#> Chain 3:                65.7565 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'estimate_infections' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000293 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 2.93 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:   1 / 700 [  0%]  (Warmup)
#> Chain 4: Iteration:  50 / 700 [  7%]  (Warmup)
#> Chain 4: Iteration: 100 / 700 [ 14%]  (Warmup)
#> Chain 4: Iteration: 150 / 700 [ 21%]  (Warmup)
#> Chain 4: Iteration: 200 / 700 [ 28%]  (Warmup)
#> Chain 4: Iteration: 201 / 700 [ 28%]  (Sampling)
#> Chain 4: Iteration: 250 / 700 [ 35%]  (Sampling)
#> Chain 4: Iteration: 300 / 700 [ 42%]  (Sampling)
#> Chain 4: Iteration: 350 / 700 [ 50%]  (Sampling)
#> Chain 4: Iteration: 400 / 700 [ 57%]  (Sampling)
#> Chain 4: Iteration: 450 / 700 [ 64%]  (Sampling)
#> Chain 4: Iteration: 500 / 700 [ 71%]  (Sampling)
#> Chain 4: Iteration: 550 / 700 [ 78%]  (Sampling)
#> Chain 4: Iteration: 600 / 700 [ 85%]  (Sampling)
#> Chain 4: Iteration: 650 / 700 [ 92%]  (Sampling)
#> Chain 4: Iteration: 700 / 700 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 27.0322 seconds (Warm-up)
#> Chain 4:                19.9045 seconds (Sampling)
#> Chain 4:                46.9367 seconds (Total)
#> Chain 4: 
#> INFO [2020-08-19 12:33:43] Completed estimates for: realland

#> Warning: There were 25 divergent transitions after warmup. See
#> http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup
#> to find out why this is a problem and how to eliminate them.

#> Warning: Examine the pairs() plot to diagnose sampling problems

# }