Summarise Regional Runtimes

Summarise Regional Runtimes

regional_runtimes(
  regional_output = NULL,
  target_folder = NULL,
  target_date = NULL,
  return_output = FALSE
)

Arguments

regional_output
A list of output as produced by regional_epinow and stored in the regional list.

target_folder
Character string specifying where to save results (will create if not present).

target_date
A character string giving the target date for which to extract results (in the format "yyyy-mm-dd"). Defaults to latest available estimates.

return_output
Logical, defaults to FALSE. Should output be returned, this automatically updates to TRUE if no directory for saving is specified.

Value

A data.table of region run times

Examples

# \donttest{
# example delays
generation_time <- get_generation_time(disease = "SARS-CoV-2", source = "ganyani")
incubation_period <- get_incubation_period(disease = "SARS-CoV-2", source = "lauer")
reporting_delay <- EpiNow2::bootstrapped_dist_fit(rlnorm(100, log(6), 1), max_value = 30)
                        
# example case vector from EpiSoon
cases <- EpiNow2::example_confirmed[1:30]
cases <- data.table::rbindlist(list(
  data.table::copy(cases)[, region := "testland"],
  cases[, region := "realland"]))
  
# run basic nowcasting pipeline
regional_out <- regional_epinow(reported_cases = cases,
                                generation_time = generation_time,
                                delays = list(incubation_period, reporting_delay),
                                samples = 100, stan_args = list(warmup = 100),
                                output = c("region", "timing"))

#> INFO [2020-11-09 20:43:14] Producing following optional outputs: regions, timing

#> Logging threshold set at INFO for the EpiNow2 logger

#> Writing EpiNow2 logs to: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmpnklt7Q/regional-epinow/2020-03-22.log

#> Logging threshold set at INFO for the EpiNow2.epinow logger

#> Writing EpiNow2.epinow logs to: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmpnklt7Q/epinow/2020-03-22.log

#> INFO [2020-11-09 20:43:14] Reporting estimates using data up to: 2020-03-22
#> INFO [2020-11-09 20:43:14] No target directory specified so returning output
#> INFO [2020-11-09 20:43:14] Producing estimates for: testland, realland
#> INFO [2020-11-09 20:43:14] Regions excluded: none
#> INFO [2020-11-09 20:43:14] Showing progress using progressr. Modify this behaviour using progressr::handlers.
#> INFO [2020-11-09 20:43:14] Completed regional estimates
#> INFO [2020-11-09 20:43:14] Regions with estimates: 0
#> INFO [2020-11-09 20:43:14] Regions with runtime errors: 2
#> INFO [2020-11-09 20:43:14] No target directory specified so returning timings

regional_runtimes(regional_output = regional_out$regional)

#> INFO [2020-11-09 20:43:14] No target directory specified so returning timings

#>      region time
#> 1: testland    0
#> 2: realland    0

# }