Report case counts by date of report

Convolves latent infections to reported cases via an observation model. Likely to be removed/replaced in later releases by functionality drawing on the stan implementation.

Usage

report_cases(
  case_estimates,
  case_forecast = NULL,
  delays,
  type = "sample",
  reporting_effect,
  CrIs = c(0.2, 0.5, 0.9)
)

Arguments

case_estimates

A data.table of case estimates with the following variables: date, sample, cases

case_forecast

A data.table of case forecasts with the following variables: date, sample, cases. If not supplied the default is not to incorporate forecasts.

delays

A call to delay_opts() defining delay distributions and options. See the documentation of delay_opts() and the examples below for details.

type

Character string indicating the method to use to transform counts. Supports either "sample" which approximates sampling or "median" would shift by the median of the distribution.

reporting_effect

A data.table giving the weekly reporting effect with the following variables: sample (must be the same as in nowcast), effect (numeric scaling factor for each weekday),day (numeric 1 - 7 (1 = Monday and 7 = Sunday)). If not supplied then no weekly reporting effect is assumed.

CrIs

Numeric vector of credible intervals to calculate.

Value

A list of data.tables. The first entry contains the following variables sample, date and cases with the second being summarised across samples.

Examples

# \donttest{
# define example cases
cases <- example_confirmed[1:40]

# get example delays
# Instead of running them model we use example
# data for speed in this example.
cases <- cases[, cases := as.integer(confirm)]
cases <- cases[, confirm := NULL][, sample := 1]

reported_cases <- report_cases(
  case_estimates = cases,
  delays = delay_opts(example_incubation_period + example_reporting_delay),
  type = "sample"
)
print(reported_cases)
#> $samples
#>     sample       date value
#>      <int>     <Date> <int>
#>  1:      1 2020-02-24     1
#>  2:      1 2020-02-25     2
#>  3:      1 2020-02-26    11
#>  4:      1 2020-02-27    14
#>  5:      1 2020-02-28    26
#>  6:      1 2020-02-29    48
#>  7:      1 2020-03-01    66
#>  8:      1 2020-03-02    76
#>  9:      1 2020-03-03   107
#> 10:      1 2020-03-04   144
#> 11:      1 2020-03-05   199
#> 12:      1 2020-03-06   216
#> 13:      1 2020-03-07   300
#> 14:      1 2020-03-08   371
#> 15:      1 2020-03-09   430
#> 16:      1 2020-03-10   483
#> 17:      1 2020-03-11   632
#> 18:      1 2020-03-12   775
#> 19:      1 2020-03-13   912
#> 20:      1 2020-03-14  1042
#> 21:      1 2020-03-15  1277
#> 22:      1 2020-03-16  1482
#> 23:      1 2020-03-17  1777
#> 24:      1 2020-03-18  2084
#> 25:      1 2020-03-19  2323
#> 26:      1 2020-03-20  2573
#> 27:      1 2020-03-21  2845
#> 28:      1 2020-03-22  3202
#> 29:      1 2020-03-23  3481
#> 30:      1 2020-03-24  3864
#> 31:      1 2020-03-25  4446
#> 32:      1 2020-03-26  4813
#> 33:      1 2020-03-27  5232
#> 34:      1 2020-03-28  5160
#> 35:      1 2020-03-29  5220
#> 36:      1 2020-03-30  5342
#> 37:      1 2020-03-31  5453
#> 38:      1 2020-04-01  5628
#>     sample       date value
#> 
#> $summarised
#>           date median  mean    sd lower_90 lower_50 lower_20 upper_20 upper_50
#>         <Date>  <num> <num> <num>    <num>    <num>    <num>    <num>    <num>
#>  1: 2020-02-24      1     1    NA        1        1        1        1        1
#>  2: 2020-02-25      2     2    NA        2        2        2        2        2
#>  3: 2020-02-26     11    11    NA       11       11       11       11       11
#>  4: 2020-02-27     14    14    NA       14       14       14       14       14
#>  5: 2020-02-28     26    26    NA       26       26       26       26       26
#>  6: 2020-02-29     48    48    NA       48       48       48       48       48
#>  7: 2020-03-01     66    66    NA       66       66       66       66       66
#>  8: 2020-03-02     76    76    NA       76       76       76       76       76
#>  9: 2020-03-03    107   107    NA      107      107      107      107      107
#> 10: 2020-03-04    144   144    NA      144      144      144      144      144
#> 11: 2020-03-05    199   199    NA      199      199      199      199      199
#> 12: 2020-03-06    216   216    NA      216      216      216      216      216
#> 13: 2020-03-07    300   300    NA      300      300      300      300      300
#> 14: 2020-03-08    371   371    NA      371      371      371      371      371
#> 15: 2020-03-09    430   430    NA      430      430      430      430      430
#> 16: 2020-03-10    483   483    NA      483      483      483      483      483
#> 17: 2020-03-11    632   632    NA      632      632      632      632      632
#> 18: 2020-03-12    775   775    NA      775      775      775      775      775
#> 19: 2020-03-13    912   912    NA      912      912      912      912      912
#> 20: 2020-03-14   1042  1042    NA     1042     1042     1042     1042     1042
#> 21: 2020-03-15   1277  1277    NA     1277     1277     1277     1277     1277
#> 22: 2020-03-16   1482  1482    NA     1482     1482     1482     1482     1482
#> 23: 2020-03-17   1777  1777    NA     1777     1777     1777     1777     1777
#> 24: 2020-03-18   2084  2084    NA     2084     2084     2084     2084     2084
#> 25: 2020-03-19   2323  2323    NA     2323     2323     2323     2323     2323
#> 26: 2020-03-20   2573  2573    NA     2573     2573     2573     2573     2573
#> 27: 2020-03-21   2845  2845    NA     2845     2845     2845     2845     2845
#> 28: 2020-03-22   3202  3202    NA     3202     3202     3202     3202     3202
#> 29: 2020-03-23   3481  3481    NA     3481     3481     3481     3481     3481
#> 30: 2020-03-24   3864  3864    NA     3864     3864     3864     3864     3864
#> 31: 2020-03-25   4446  4446    NA     4446     4446     4446     4446     4446
#> 32: 2020-03-26   4813  4813    NA     4813     4813     4813     4813     4813
#> 33: 2020-03-27   5232  5232    NA     5232     5232     5232     5232     5232
#> 34: 2020-03-28   5160  5160    NA     5160     5160     5160     5160     5160
#> 35: 2020-03-29   5220  5220    NA     5220     5220     5220     5220     5220
#> 36: 2020-03-30   5342  5342    NA     5342     5342     5342     5342     5342
#> 37: 2020-03-31   5453  5453    NA     5453     5453     5453     5453     5453
#> 38: 2020-04-01   5628  5628    NA     5628     5628     5628     5628     5628
#>           date median  mean    sd lower_90 lower_50 lower_20 upper_20 upper_50
#>     upper_90
#>        <num>
#>  1:        1
#>  2:        2
#>  3:       11
#>  4:       14
#>  5:       26
#>  6:       48
#>  7:       66
#>  8:       76
#>  9:      107
#> 10:      144
#> 11:      199
#> 12:      216
#> 13:      300
#> 14:      371
#> 15:      430
#> 16:      483
#> 17:      632
#> 18:      775
#> 19:      912
#> 20:     1042
#> 21:     1277
#> 22:     1482
#> 23:     1777
#> 24:     2084
#> 25:     2323
#> 26:     2573
#> 27:     2845
#> 28:     3202
#> 29:     3481
#> 30:     3864
#> 31:     4446
#> 32:     4813
#> 33:     5232
#> 34:     5160
#> 35:     5220
#> 36:     5342
#> 37:     5453
#> 38:     5628
#>     upper_90
#> 
# }