Report case counts by date of report

Convolves latent infections to reported cases via an observation model. Likely to be removed/replaced in later releases by functionality drawing on the stan implementation.

Usage

report_cases(
  case_estimates,
  case_forecast = NULL,
  delays,
  type = "sample",
  reporting_effect,
  CrIs = c(0.2, 0.5, 0.9)
)

Arguments

case_estimates

A data.table of case estimates with the following variables: date, sample, cases

case_forecast

A data.table of case forecasts with the following variables: date, sample, cases. If not supplied the default is not to incorporate forecasts.

delays

A call to delay_opts() defining delay distributions and options. See the documentation of delay_opts() and the examples below for details.

type

Character string indicating the method to use to transform counts. Supports either "sample" which approximates sampling or "median" would shift by the median of the distribution.

reporting_effect

A data.table giving the weekly reporting effect with the following variables: sample (must be the same as in nowcast), effect (numeric scaling factor for each weekday),day (numeric 1 - 7 (1 = Monday and 7 = Sunday)). If not supplied then no weekly reporting effect is assumed.

CrIs

Numeric vector of credible intervals to calculate.

Value

A list of data.tables. The first entry contains the following variables sample, date and cases with the second being summarised across samples.

Examples

# \donttest{
# define example cases
cases <- example_confirmed[1:40]

# get example delays
# Instead of running them model we use example
# data for speed in this example.
cases <- cases[, cases := as.integer(confirm)]
cases <- cases[, confirm := NULL][, sample := 1]

reported_cases <- report_cases(
  case_estimates = cases,
  delays = delay_opts(example_incubation_period + example_reporting_delay),
  type = "sample"
)
print(reported_cases)
#> $samples
#>     sample       date value
#>      <int>     <Date> <int>
#>  1:      1 2020-02-22     3
#>  2:      1 2020-02-23    18
#>  3:      1 2020-02-24    24
#>  4:      1 2020-02-25    41
#>  5:      1 2020-02-26    52
#>  6:      1 2020-02-27    63
#>  7:      1 2020-02-28   104
#>  8:      1 2020-02-29   129
#>  9:      1 2020-03-01   152
#> 10:      1 2020-03-02   240
#> 11:      1 2020-03-03   269
#> 12:      1 2020-03-04   285
#> 13:      1 2020-03-05   389
#> 14:      1 2020-03-06   474
#> 15:      1 2020-03-07   527
#> 16:      1 2020-03-08   683
#> 17:      1 2020-03-09   895
#> 18:      1 2020-03-10  1067
#> 19:      1 2020-03-11   971
#> 20:      1 2020-03-12  1306
#> 21:      1 2020-03-13  1583
#> 22:      1 2020-03-14  1750
#> 23:      1 2020-03-15  2122
#> 24:      1 2020-03-16  2263
#> 25:      1 2020-03-17  2612
#> 26:      1 2020-03-18  2849
#> 27:      1 2020-03-19  3140
#> 28:      1 2020-03-20  3563
#> 29:      1 2020-03-21  4077
#> 30:      1 2020-03-22  4564
#> 31:      1 2020-03-23  4669
#> 32:      1 2020-03-24  4675
#> 33:      1 2020-03-25  4884
#> 34:      1 2020-03-26  5070
#> 35:      1 2020-03-27  5265
#> 36:      1 2020-03-28  5257
#> 37:      1 2020-03-29  5525
#> 38:      1 2020-03-30  5424
#> 39:      1 2020-03-31  5193
#> 40:      1 2020-04-01  4933
#>     sample       date value
#> 
#> $summarised
#>           date median  mean    sd lower_90 lower_50 lower_20 upper_20 upper_50
#>         <Date>  <num> <num> <num>    <num>    <num>    <num>    <num>    <num>
#>  1: 2020-02-22      3     3    NA        3        3        3        3        3
#>  2: 2020-02-23     18    18    NA       18       18       18       18       18
#>  3: 2020-02-24     24    24    NA       24       24       24       24       24
#>  4: 2020-02-25     41    41    NA       41       41       41       41       41
#>  5: 2020-02-26     52    52    NA       52       52       52       52       52
#>  6: 2020-02-27     63    63    NA       63       63       63       63       63
#>  7: 2020-02-28    104   104    NA      104      104      104      104      104
#>  8: 2020-02-29    129   129    NA      129      129      129      129      129
#>  9: 2020-03-01    152   152    NA      152      152      152      152      152
#> 10: 2020-03-02    240   240    NA      240      240      240      240      240
#> 11: 2020-03-03    269   269    NA      269      269      269      269      269
#> 12: 2020-03-04    285   285    NA      285      285      285      285      285
#> 13: 2020-03-05    389   389    NA      389      389      389      389      389
#> 14: 2020-03-06    474   474    NA      474      474      474      474      474
#> 15: 2020-03-07    527   527    NA      527      527      527      527      527
#> 16: 2020-03-08    683   683    NA      683      683      683      683      683
#> 17: 2020-03-09    895   895    NA      895      895      895      895      895
#> 18: 2020-03-10   1067  1067    NA     1067     1067     1067     1067     1067
#> 19: 2020-03-11    971   971    NA      971      971      971      971      971
#> 20: 2020-03-12   1306  1306    NA     1306     1306     1306     1306     1306
#> 21: 2020-03-13   1583  1583    NA     1583     1583     1583     1583     1583
#> 22: 2020-03-14   1750  1750    NA     1750     1750     1750     1750     1750
#> 23: 2020-03-15   2122  2122    NA     2122     2122     2122     2122     2122
#> 24: 2020-03-16   2263  2263    NA     2263     2263     2263     2263     2263
#> 25: 2020-03-17   2612  2612    NA     2612     2612     2612     2612     2612
#> 26: 2020-03-18   2849  2849    NA     2849     2849     2849     2849     2849
#> 27: 2020-03-19   3140  3140    NA     3140     3140     3140     3140     3140
#> 28: 2020-03-20   3563  3563    NA     3563     3563     3563     3563     3563
#> 29: 2020-03-21   4077  4077    NA     4077     4077     4077     4077     4077
#> 30: 2020-03-22   4564  4564    NA     4564     4564     4564     4564     4564
#> 31: 2020-03-23   4669  4669    NA     4669     4669     4669     4669     4669
#> 32: 2020-03-24   4675  4675    NA     4675     4675     4675     4675     4675
#> 33: 2020-03-25   4884  4884    NA     4884     4884     4884     4884     4884
#> 34: 2020-03-26   5070  5070    NA     5070     5070     5070     5070     5070
#> 35: 2020-03-27   5265  5265    NA     5265     5265     5265     5265     5265
#> 36: 2020-03-28   5257  5257    NA     5257     5257     5257     5257     5257
#> 37: 2020-03-29   5525  5525    NA     5525     5525     5525     5525     5525
#> 38: 2020-03-30   5424  5424    NA     5424     5424     5424     5424     5424
#> 39: 2020-03-31   5193  5193    NA     5193     5193     5193     5193     5193
#> 40: 2020-04-01   4933  4933    NA     4933     4933     4933     4933     4933
#>           date median  mean    sd lower_90 lower_50 lower_20 upper_20 upper_50
#>     upper_90
#>        <num>
#>  1:        3
#>  2:       18
#>  3:       24
#>  4:       41
#>  5:       52
#>  6:       63
#>  7:      104
#>  8:      129
#>  9:      152
#> 10:      240
#> 11:      269
#> 12:      285
#> 13:      389
#> 14:      474
#> 15:      527
#> 16:      683
#> 17:      895
#> 18:     1067
#> 19:      971
#> 20:     1306
#> 21:     1583
#> 22:     1750
#> 23:     2122
#> 24:     2263
#> 25:     2612
#> 26:     2849
#> 27:     3140
#> 28:     3563
#> 29:     4077
#> 30:     4564
#> 31:     4669
#> 32:     4675
#> 33:     4884
#> 34:     5070
#> 35:     5265
#> 36:     5257
#> 37:     5525
#> 38:     5424
#> 39:     5193
#> 40:     4933
#>     upper_90
#> 
# }