Report case counts by date of report

Convolves latent infections to reported cases via an observation model. Likely to be removed/replaced in later releases by functionality drawing on the stan implementation.

Usage

report_cases(
  case_estimates,
  case_forecast = NULL,
  delays,
  type = "sample",
  reporting_effect,
  CrIs = c(0.2, 0.5, 0.9)
)

Arguments

case_estimates

A data.table of case estimates with the following variables: date, sample, cases

case_forecast

A data.table of case forecasts with the following variables: date, sample, cases. If not supplied the default is not to incorporate forecasts.

delays

A call to delay_opts() defining delay distributions and options. See the documentation of delay_opts() and the examples below for details.

type

Character string indicating the method to use to transform counts. Supports either "sample" which approximates sampling or "median" would shift by the median of the distribution.

reporting_effect

A data.table giving the weekly reporting effect with the following variables: sample (must be the same as in nowcast), effect (numeric scaling factor for each weekday),day (numeric 1 - 7 (1 = Monday and 7 = Sunday)). If not supplied then no weekly reporting effect is assumed.

CrIs

Numeric vector of credible intervals to calculate.

Value

A list of data.tables. The first entry contains the following variables sample, date and cases with the second being summarised across samples.

Examples

# \donttest{
# define example cases
cases <- example_confirmed[1:40]

# get example delays
# Instead of running them model we use example
# data for speed in this example.
cases <- cases[, cases := as.integer(confirm)]
cases <- cases[, confirm := NULL][, sample := 1]

reported_cases <- report_cases(
  case_estimates = cases,
  delays = delay_opts(example_incubation_period + example_reporting_delay),
  type = "sample"
)
print(reported_cases)
#> $samples
#>     sample       date value
#>      <int>     <Date> <int>
#>  1:      1 2020-02-22     3
#>  2:      1 2020-02-23    18
#>  3:      1 2020-02-24    30
#>  4:      1 2020-02-25    38
#>  5:      1 2020-02-26    54
#>  6:      1 2020-02-27    68
#>  7:      1 2020-02-28   104
#>  8:      1 2020-02-29   121
#>  9:      1 2020-03-01   157
#> 10:      1 2020-03-02   232
#> 11:      1 2020-03-03   257
#> 12:      1 2020-03-04   298
#> 13:      1 2020-03-05   348
#> 14:      1 2020-03-06   443
#> 15:      1 2020-03-07   481
#> 16:      1 2020-03-08   665
#> 17:      1 2020-03-09   886
#> 18:      1 2020-03-10  1111
#> 19:      1 2020-03-11  1020
#> 20:      1 2020-03-12  1316
#> 21:      1 2020-03-13  1658
#> 22:      1 2020-03-14  1791
#> 23:      1 2020-03-15  2158
#> 24:      1 2020-03-16  2320
#> 25:      1 2020-03-17  2661
#> 26:      1 2020-03-18  2801
#> 27:      1 2020-03-19  3104
#> 28:      1 2020-03-20  3664
#> 29:      1 2020-03-21  4188
#> 30:      1 2020-03-22  4721
#> 31:      1 2020-03-23  4751
#> 32:      1 2020-03-24  4814
#> 33:      1 2020-03-25  4809
#> 34:      1 2020-03-26  4959
#> 35:      1 2020-03-27  5204
#> 36:      1 2020-03-28  5406
#> 37:      1 2020-03-29  5572
#> 38:      1 2020-03-30  5489
#> 39:      1 2020-03-31  5027
#> 40:      1 2020-04-01  4822
#>     sample       date value
#> 
#> $summarised
#>           date median  mean    sd lower_90 lower_50 lower_20 upper_20 upper_50
#>         <Date>  <num> <num> <num>    <num>    <num>    <num>    <num>    <num>
#>  1: 2020-02-22      3     3    NA        3        3        3        3        3
#>  2: 2020-02-23     18    18    NA       18       18       18       18       18
#>  3: 2020-02-24     30    30    NA       30       30       30       30       30
#>  4: 2020-02-25     38    38    NA       38       38       38       38       38
#>  5: 2020-02-26     54    54    NA       54       54       54       54       54
#>  6: 2020-02-27     68    68    NA       68       68       68       68       68
#>  7: 2020-02-28    104   104    NA      104      104      104      104      104
#>  8: 2020-02-29    121   121    NA      121      121      121      121      121
#>  9: 2020-03-01    157   157    NA      157      157      157      157      157
#> 10: 2020-03-02    232   232    NA      232      232      232      232      232
#> 11: 2020-03-03    257   257    NA      257      257      257      257      257
#> 12: 2020-03-04    298   298    NA      298      298      298      298      298
#> 13: 2020-03-05    348   348    NA      348      348      348      348      348
#> 14: 2020-03-06    443   443    NA      443      443      443      443      443
#> 15: 2020-03-07    481   481    NA      481      481      481      481      481
#> 16: 2020-03-08    665   665    NA      665      665      665      665      665
#> 17: 2020-03-09    886   886    NA      886      886      886      886      886
#> 18: 2020-03-10   1111  1111    NA     1111     1111     1111     1111     1111
#> 19: 2020-03-11   1020  1020    NA     1020     1020     1020     1020     1020
#> 20: 2020-03-12   1316  1316    NA     1316     1316     1316     1316     1316
#> 21: 2020-03-13   1658  1658    NA     1658     1658     1658     1658     1658
#> 22: 2020-03-14   1791  1791    NA     1791     1791     1791     1791     1791
#> 23: 2020-03-15   2158  2158    NA     2158     2158     2158     2158     2158
#> 24: 2020-03-16   2320  2320    NA     2320     2320     2320     2320     2320
#> 25: 2020-03-17   2661  2661    NA     2661     2661     2661     2661     2661
#> 26: 2020-03-18   2801  2801    NA     2801     2801     2801     2801     2801
#> 27: 2020-03-19   3104  3104    NA     3104     3104     3104     3104     3104
#> 28: 2020-03-20   3664  3664    NA     3664     3664     3664     3664     3664
#> 29: 2020-03-21   4188  4188    NA     4188     4188     4188     4188     4188
#> 30: 2020-03-22   4721  4721    NA     4721     4721     4721     4721     4721
#> 31: 2020-03-23   4751  4751    NA     4751     4751     4751     4751     4751
#> 32: 2020-03-24   4814  4814    NA     4814     4814     4814     4814     4814
#> 33: 2020-03-25   4809  4809    NA     4809     4809     4809     4809     4809
#> 34: 2020-03-26   4959  4959    NA     4959     4959     4959     4959     4959
#> 35: 2020-03-27   5204  5204    NA     5204     5204     5204     5204     5204
#> 36: 2020-03-28   5406  5406    NA     5406     5406     5406     5406     5406
#> 37: 2020-03-29   5572  5572    NA     5572     5572     5572     5572     5572
#> 38: 2020-03-30   5489  5489    NA     5489     5489     5489     5489     5489
#> 39: 2020-03-31   5027  5027    NA     5027     5027     5027     5027     5027
#> 40: 2020-04-01   4822  4822    NA     4822     4822     4822     4822     4822
#>           date median  mean    sd lower_90 lower_50 lower_20 upper_20 upper_50
#>     upper_90
#>        <num>
#>  1:        3
#>  2:       18
#>  3:       30
#>  4:       38
#>  5:       54
#>  6:       68
#>  7:      104
#>  8:      121
#>  9:      157
#> 10:      232
#> 11:      257
#> 12:      298
#> 13:      348
#> 14:      443
#> 15:      481
#> 16:      665
#> 17:      886
#> 18:     1111
#> 19:     1020
#> 20:     1316
#> 21:     1658
#> 22:     1791
#> 23:     2158
#> 24:     2320
#> 25:     2661
#> 26:     2801
#> 27:     3104
#> 28:     3664
#> 29:     4188
#> 30:     4721
#> 31:     4751
#> 32:     4814
#> 33:     4809
#> 34:     4959
#> 35:     5204
#> 36:     5406
#> 37:     5572
#> 38:     5489
#> 39:     5027
#> 40:     4822
#>     upper_90
#> 
# }