Report case counts by date of report

Convolves latent infections to reported cases via an observation model. Likely to be removed/replaced in later releases by functionality drawing on the stan implementation.

Usage

report_cases(
  case_estimates,
  case_forecast = NULL,
  delays,
  type = "sample",
  reporting_effect,
  CrIs = c(0.2, 0.5, 0.9)
)

Arguments

case_estimates

A data.table of case estimates with the following variables: date, sample, cases

case_forecast

A data.table of case forecasts with the following variables: date, sample, cases. If not supplied the default is not to incorporate forecasts.

delays

A call to delay_opts() defining delay distributions and options. See the documentation of delay_opts() and the examples below for details.

type

Character string indicating the method to use to transform counts. Supports either "sample" which approximates sampling or "median" would shift by the median of the distribution.

reporting_effect

A data.table giving the weekly reporting effect with the following variables: sample (must be the same as in nowcast), effect (numeric scaling factor for each weekday),day (numeric 1 - 7 (1 = Monday and 7 = Sunday)). If not supplied then no weekly reporting effect is assumed.

CrIs

Numeric vector of credible intervals to calculate.

Value

A list of data.tables. The first entry contains the following variables sample, date and cases with the second being summarised across samples.

Examples

# \donttest{
# define example cases
cases <- example_confirmed[1:40]

# set up example delays
generation_time <- get_generation_time(
 disease = "SARS-CoV-2", source = "ganyani"
)
incubation_period <- get_incubation_period(
 disease = "SARS-CoV-2", source = "lauer"
)
reporting_delay <- list(
  mean = convert_to_logmean(2, 1), mean_sd = 0.1,
  sd = convert_to_logsd(2, 1), sd_sd = 0.1, max = 10
)

# Instead of running them model we use example
# data for speed in this example.
cases <- cases[, cases := as.integer(confirm)]
cases <- cases[, confirm := NULL][, sample := 1]

reported_cases <- report_cases(
  case_estimates = cases,
  delays = delay_opts(incubation_period, reporting_delay),
  type = "sample"
)
print(reported_cases)
#> $samples
#>     sample       date value
#>  1:      1 2020-02-23     2
#>  2:      1 2020-02-24     6
#>  3:      1 2020-02-25    14
#>  4:      1 2020-02-26    23
#>  5:      1 2020-02-27    30
#>  6:      1 2020-02-28    45
#>  7:      1 2020-02-29    67
#>  8:      1 2020-03-01    88
#>  9:      1 2020-03-02   116
#> 10:      1 2020-03-03   142
#> 11:      1 2020-03-04   216
#> 12:      1 2020-03-05   242
#> 13:      1 2020-03-06   312
#> 14:      1 2020-03-07   333
#> 15:      1 2020-03-08   389
#> 16:      1 2020-03-09   507
#> 17:      1 2020-03-10   648
#> 18:      1 2020-03-11   732
#> 19:      1 2020-03-12   947
#> 20:      1 2020-03-13  1024
#> 21:      1 2020-03-14  1244
#> 22:      1 2020-03-15  1447
#> 23:      1 2020-03-16  1715
#> 24:      1 2020-03-17  1872
#> 25:      1 2020-03-18  2118
#> 26:      1 2020-03-19  2410
#> 27:      1 2020-03-20  2562
#> 28:      1 2020-03-21  2901
#> 29:      1 2020-03-22  3261
#> 30:      1 2020-03-23  3757
#> 31:      1 2020-03-24  3972
#> 32:      1 2020-03-25  4371
#> 33:      1 2020-03-26  4417
#> 34:      1 2020-03-27  4492
#> 35:      1 2020-03-28  4503
#> 36:      1 2020-03-29  4598
#> 37:      1 2020-03-30  4626
#> 38:      1 2020-03-31  4790
#> 39:      1 2020-04-01  4761
#>     sample       date value
#> 
#> $summarised
#>           date median mean sd lower_90 lower_50 lower_20 upper_20 upper_50
#>  1: 2020-02-23      2    2 NA        2        2        2        2        2
#>  2: 2020-02-24      6    6 NA        6        6        6        6        6
#>  3: 2020-02-25     14   14 NA       14       14       14       14       14
#>  4: 2020-02-26     23   23 NA       23       23       23       23       23
#>  5: 2020-02-27     30   30 NA       30       30       30       30       30
#>  6: 2020-02-28     45   45 NA       45       45       45       45       45
#>  7: 2020-02-29     67   67 NA       67       67       67       67       67
#>  8: 2020-03-01     88   88 NA       88       88       88       88       88
#>  9: 2020-03-02    116  116 NA      116      116      116      116      116
#> 10: 2020-03-03    142  142 NA      142      142      142      142      142
#> 11: 2020-03-04    216  216 NA      216      216      216      216      216
#> 12: 2020-03-05    242  242 NA      242      242      242      242      242
#> 13: 2020-03-06    312  312 NA      312      312      312      312      312
#> 14: 2020-03-07    333  333 NA      333      333      333      333      333
#> 15: 2020-03-08    389  389 NA      389      389      389      389      389
#> 16: 2020-03-09    507  507 NA      507      507      507      507      507
#> 17: 2020-03-10    648  648 NA      648      648      648      648      648
#> 18: 2020-03-11    732  732 NA      732      732      732      732      732
#> 19: 2020-03-12    947  947 NA      947      947      947      947      947
#> 20: 2020-03-13   1024 1024 NA     1024     1024     1024     1024     1024
#> 21: 2020-03-14   1244 1244 NA     1244     1244     1244     1244     1244
#> 22: 2020-03-15   1447 1447 NA     1447     1447     1447     1447     1447
#> 23: 2020-03-16   1715 1715 NA     1715     1715     1715     1715     1715
#> 24: 2020-03-17   1872 1872 NA     1872     1872     1872     1872     1872
#> 25: 2020-03-18   2118 2118 NA     2118     2118     2118     2118     2118
#> 26: 2020-03-19   2410 2410 NA     2410     2410     2410     2410     2410
#> 27: 2020-03-20   2562 2562 NA     2562     2562     2562     2562     2562
#> 28: 2020-03-21   2901 2901 NA     2901     2901     2901     2901     2901
#> 29: 2020-03-22   3261 3261 NA     3261     3261     3261     3261     3261
#> 30: 2020-03-23   3757 3757 NA     3757     3757     3757     3757     3757
#> 31: 2020-03-24   3972 3972 NA     3972     3972     3972     3972     3972
#> 32: 2020-03-25   4371 4371 NA     4371     4371     4371     4371     4371
#> 33: 2020-03-26   4417 4417 NA     4417     4417     4417     4417     4417
#> 34: 2020-03-27   4492 4492 NA     4492     4492     4492     4492     4492
#> 35: 2020-03-28   4503 4503 NA     4503     4503     4503     4503     4503
#> 36: 2020-03-29   4598 4598 NA     4598     4598     4598     4598     4598
#> 37: 2020-03-30   4626 4626 NA     4626     4626     4626     4626     4626
#> 38: 2020-03-31   4790 4790 NA     4790     4790     4790     4790     4790
#> 39: 2020-04-01   4761 4761 NA     4761     4761     4761     4761     4761
#>           date median mean sd lower_90 lower_50 lower_20 upper_20 upper_50
#>     upper_90
#>  1:        2
#>  2:        6
#>  3:       14
#>  4:       23
#>  5:       30
#>  6:       45
#>  7:       67
#>  8:       88
#>  9:      116
#> 10:      142
#> 11:      216
#> 12:      242
#> 13:      312
#> 14:      333
#> 15:      389
#> 16:      507
#> 17:      648
#> 18:      732
#> 19:      947
#> 20:     1024
#> 21:     1244
#> 22:     1447
#> 23:     1715
#> 24:     1872
#> 25:     2118
#> 26:     2410
#> 27:     2562
#> 28:     2901
#> 29:     3261
#> 30:     3757
#> 31:     3972
#> 32:     4371
#> 33:     4417
#> 34:     4492
#> 35:     4503
#> 36:     4598
#> 37:     4626
#> 38:     4790
#> 39:     4761
#>     upper_90
#> 
# }