Report case counts by date of report

Report case counts by date of report

report_cases(
  case_estimates,
  case_forecast = NULL,
  delays,
  type = "sample",
  reporting_effect
)

Arguments

case_estimates
A data.table of case estimates with the following variables: date, sample, cases

case_forecast
A data.table of case forecasts with the following variables: date, sample, cases. If not supplied the default is not to incorporate forecasts.

delays
A list of delays (i.e incubation period/reporting delay) between infection and report. Each list entry must also be a list containing the mean, standard deviation of the mean (mean_sd), standard deviation (sd), standard deviation of the standard deviation and the maximum allowed value for the that delay (assuming a lognormal distribution with all parameters excepting the max allowed value on the log scale).

type
Character string indicating the method to use to transform counts. Supports either "sample" which approximates sampling or "median" would shift by the median of the distribution.

reporting_effect
A data.table giving the weekly reporting effect with the following variables: sample (must be the same as in nowcast), effect (numeric scaling factor for each weekday), day (numeric 1 - 7 (1 = Monday and 7 = Sunday)). If not supplied then no weekly reporting effect is assumed.

Value

A list of data.tables. The first entry contains the following variables sample, date and cases with the second being summarised across samples.

Examples

# \donttest{
## Define example cases
cases <- EpiNow2::example_confirmed[1:40]

 
## Set up example generation time
generation_time <- list(mean = EpiNow2::covid_generation_times[1, ]$mean,
                        mean_sd = EpiNow2::covid_generation_times[1, ]$mean_sd,
                        sd = EpiNow2::covid_generation_times[1, ]$sd,
                        sd_sd = EpiNow2::covid_generation_times[1, ]$sd_sd,
                        max = 30)
## Set                   
incubation_period <- list(mean = EpiNow2::covid_incubation_period[1, ]$mean,
                          mean_sd = EpiNow2::covid_incubation_period[1, ]$mean_sd,
                          sd = EpiNow2::covid_incubation_period[1, ]$sd,
                          sd_sd = EpiNow2::covid_incubation_period[1, ]$sd_sd,
                          max = 30)
                   
reporting_delay <- list(mean = log(5),
                        mean_sd = log(2),
                        sd = log(2),
                        sd_sd = log(1.5),
                        max = 30)
                        
## Run model
out <- EpiNow2::estimate_infections(cases, family = "negbin",
                                    generation_time = generation_time,
                                    delays = list(incubation_period, reporting_delay),
                                    samples = 1000, warmup = 200, 
                                    cores = ifelse(interactive(), 4, 1), chains = 4,
                                    estimate_rt =  FALSE, verbose = TRUE)

#> Running for 1000 samples (across 4 chains each with a warm up of 200 iterations each) and 58 time steps of which 7 are a forecast

#> 
#> SAMPLING FOR MODEL 'estimate_infections' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000168 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.68 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 450 [  0%]  (Warmup)
#> Chain 1: Iteration:  50 / 450 [ 11%]  (Warmup)
#> Chain 1: Iteration: 100 / 450 [ 22%]  (Warmup)
#> Chain 1: Iteration: 150 / 450 [ 33%]  (Warmup)
#> Chain 1: Iteration: 200 / 450 [ 44%]  (Warmup)
#> Chain 1: Iteration: 201 / 450 [ 44%]  (Sampling)
#> Chain 1: Iteration: 250 / 450 [ 55%]  (Sampling)
#> Chain 1: Iteration: 300 / 450 [ 66%]  (Sampling)
#> Chain 1: Iteration: 350 / 450 [ 77%]  (Sampling)
#> Chain 1: Iteration: 400 / 450 [ 88%]  (Sampling)
#> Chain 1: Iteration: 450 / 450 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 32.1431 seconds (Warm-up)
#> Chain 1:                1.75172 seconds (Sampling)
#> Chain 1:                33.8948 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'estimate_infections' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000405 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 4.05 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:   1 / 450 [  0%]  (Warmup)
#> Chain 2: Iteration:  50 / 450 [ 11%]  (Warmup)
#> Chain 2: Iteration: 100 / 450 [ 22%]  (Warmup)
#> Chain 2: Iteration: 150 / 450 [ 33%]  (Warmup)
#> Chain 2: Iteration: 200 / 450 [ 44%]  (Warmup)
#> Chain 2: Iteration: 201 / 450 [ 44%]  (Sampling)
#> Chain 2: Iteration: 250 / 450 [ 55%]  (Sampling)
#> Chain 2: Iteration: 300 / 450 [ 66%]  (Sampling)
#> Chain 2: Iteration: 350 / 450 [ 77%]  (Sampling)
#> Chain 2: Iteration: 400 / 450 [ 88%]  (Sampling)
#> Chain 2: Iteration: 450 / 450 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 35.0007 seconds (Warm-up)
#> Chain 2:                2.04112 seconds (Sampling)
#> Chain 2:                37.0418 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'estimate_infections' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000417 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 4.17 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:   1 / 450 [  0%]  (Warmup)
#> Chain 3: Iteration:  50 / 450 [ 11%]  (Warmup)
#> Chain 3: Iteration: 100 / 450 [ 22%]  (Warmup)
#> Chain 3: Iteration: 150 / 450 [ 33%]  (Warmup)
#> Chain 3: Iteration: 200 / 450 [ 44%]  (Warmup)
#> Chain 3: Iteration: 201 / 450 [ 44%]  (Sampling)
#> Chain 3: Iteration: 250 / 450 [ 55%]  (Sampling)
#> Chain 3: Iteration: 300 / 450 [ 66%]  (Sampling)
#> Chain 3: Iteration: 350 / 450 [ 77%]  (Sampling)
#> Chain 3: Iteration: 400 / 450 [ 88%]  (Sampling)
#> Chain 3: Iteration: 450 / 450 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 26.6697 seconds (Warm-up)
#> Chain 3:                1.82832 seconds (Sampling)
#> Chain 3:                28.498 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'estimate_infections' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000149 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.49 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:   1 / 450 [  0%]  (Warmup)
#> Chain 4: Iteration:  50 / 450 [ 11%]  (Warmup)
#> Chain 4: Iteration: 100 / 450 [ 22%]  (Warmup)
#> Chain 4: Iteration: 150 / 450 [ 33%]  (Warmup)
#> Chain 4: Iteration: 200 / 450 [ 44%]  (Warmup)
#> Chain 4: Iteration: 201 / 450 [ 44%]  (Sampling)
#> Chain 4: Iteration: 250 / 450 [ 55%]  (Sampling)
#> Chain 4: Iteration: 300 / 450 [ 66%]  (Sampling)
#> Chain 4: Iteration: 350 / 450 [ 77%]  (Sampling)
#> Chain 4: Iteration: 400 / 450 [ 88%]  (Sampling)
#> Chain 4: Iteration: 450 / 450 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 27.7043 seconds (Warm-up)
#> Chain 4:                2.07244 seconds (Sampling)
#> Chain 4:                29.7767 seconds (Total)
#> Chain 4: 

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

                            
                     
reported_cases <- report_cases(case_estimates = out$samples[variable == "infections"][, 
                                                            cases := value][, value := NULL],
                               delays = list(incubation_period, reporting_delay),
                               type = "sample")

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Warning: NAs produced

#> Error in rep(0, length(reversed_cases) - .): invalid 'times' argument

                               
print(reported_cases)

#> Error in print(reported_cases): object 'reported_cases' not found

# }