Truncation Distribution Options

Returns a truncation distribution formatted for usage by downstream functions. See estimate_truncation for an approach to estimate this distribution.

trunc_opts(dist = NULL)

Arguments

dist
A list defining the truncation distribution, defaults to NULL in which case no truncation is used. Must have the following elements if defined: "mean", "mean_sd", "sd_mean", "sd_sd", and "max" defining a truncated log normal (with all parameters except for max defined in logged form).

Value

A list summarising the input truncation distribution.

See also

Examples

# no truncation
trunc_opts()

#> $truncation
#> [1] 0
#> 
#> $trunc_mean_mean
#> numeric(0)
#> 
#> $trunc_mean_sd
#> numeric(0)
#> 
#> $trunc_sd_mean
#> numeric(0)
#> 
#> $trunc_sd_sd
#> numeric(0)
#> 
#> $max_truncation
#> numeric(0)
#>